+++ src/main.py
@@ -5,9 +5,10 @@ sys.path.insert(0, os.path.dirname(os.path.dirname(__file__)))
 
 from flask import Flask, send_from_directory
 from flask_cors import CORS
-from src.models.user import db
-from src.routes.user import user_bp
+
 from src.routes.golden_plate_recorder_db import recorder_bp
+from src.routes.golden_plate_recorder_db.db import db_session
+
 
 app = Flask(__name__, static_folder=os.path.join(os.path.dirname(__file__), 'static'))
 app.config['SECRET_KEY'] = 'asdf#FGSgvasgf$5$WGT'
@@ -15,15 +16,7 @@ app.config['SECRET_KEY'] = 'asdf#FGSgvasgf$5$WGT'
 # Enable CORS for all routes
 CORS(app, supports_credentials=True)
 
-app.register_blueprint(user_bp, url_prefix='/api')
 app.register_blueprint(recorder_bp, url_prefix='/api')
-
-# uncomment if you need to use database
-app.config['SQLALCHEMY_DATABASE_URI'] = f"sqlite:///{os.path.join(os.path.dirname(__file__), 'database', 'app.db')}"
-app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
-db.init_app(app)
-with app.app_context():
-    db.create_all()
 
 @app.route('/', defaults={'path': ''})
 @app.route('/<path:path>')
@@ -40,6 +33,11 @@ def serve(path):
             return send_from_directory(static_folder_path, 'index.html')
         else:
             return "index.html not found", 404
+
+
+@app.teardown_appcontext
+def _shutdown_scoped_session(exception=None):
+    db_session.remove()
 
 
 if __name__ == '__main__':
