--- src/routes/golden_plate_recorder_db.py
@@ -1,3139 +1,0 @@
-from flask import Blueprint, request, jsonify, session
-from datetime import datetime, timezone
-import uuid
-import csv
-import io
-import os
-import json
-import random
-import re
-
-from sqlalchemy import (
-    create_engine,
-    Column,
-    String,
-    Text,
-    DateTime,
-    ForeignKey,
-    Index,
-)
-from sqlalchemy.orm import declarative_base, sessionmaker, scoped_session
-
-recorder_bp = Blueprint('recorder', __name__)
-
-DATABASE_URL = os.environ.get('DATABASE_URL', 'sqlite:///data/golden_plate_recorder.db')
-if DATABASE_URL.startswith('sqlite:///'):
-    db_path = DATABASE_URL.replace('sqlite:///', '', 1)
-    db_dir = os.path.dirname(db_path)
-    if db_dir:
-        os.makedirs(db_dir, exist_ok=True)
-
-engine = create_engine(
-    DATABASE_URL,
-    connect_args={'check_same_thread': False} if DATABASE_URL.startswith('sqlite') else {}
-)
-SessionFactory = sessionmaker(autocommit=False, autoflush=False, bind=engine)
-db_session = scoped_session(SessionFactory)
-Base = declarative_base()
-
-
-def _now_utc():
-    return datetime.now(timezone.utc)
-
-
-class KeyValueStore(Base):
-    __tablename__ = 'kv_store'
-
-    key = Column(String(128), primary_key=True)
-    value = Column(Text, nullable=False, default='{}')
-    updated_at = Column(DateTime(timezone=True), default=_now_utc, onupdate=_now_utc)
-
-
-class User(Base):
-    __tablename__ = 'users'
-    __table_args__ = (Index('idx_users_username', 'username'),)
-
-    id = Column(String, primary_key=True, default=lambda: str(uuid.uuid4()))
-    username = Column(String, unique=True, nullable=False)
-    password_hash = Column(Text, nullable=False)
-    display_name = Column(String, nullable=False)
-    role = Column(String, nullable=False)
-    last_login_at = Column(DateTime(timezone=True))
-    created_at = Column(DateTime(timezone=True), default=_now_utc)
-    updated_at = Column(DateTime(timezone=True), default=_now_utc, onupdate=_now_utc)
-    status = Column(String, nullable=False, default='active')
-
-
-class UserInviteCode(Base):
-    __tablename__ = 'user_invite_codes'
-    __table_args__ = (Index('idx_invite_codes_user', 'user_id', 'status'),)
-
-    id = Column(String, primary_key=True, default=lambda: str(uuid.uuid4()))
-    user_id = Column(String, ForeignKey('users.id', ondelete='CASCADE'), nullable=False)
-    code = Column(String, unique=True, nullable=False)
-    issued_by = Column(String, ForeignKey('users.id'), nullable=False)
-    issued_at = Column(DateTime(timezone=True), default=_now_utc)
-    expires_at = Column(DateTime(timezone=True))
-    used_by = Column(String, ForeignKey('users.id'))
-    used_at = Column(DateTime(timezone=True))
-    status = Column(String, nullable=False, default='unused')
-    role = Column(String, nullable=False, default='user')
-
-
-class Student(Base):
-    __tablename__ = 'students'
-
-    id = Column(String, primary_key=True, default=lambda: str(uuid.uuid4()))
-    student_identifier = Column(String, unique=True, nullable=False)
-    preferred_name = Column(String, nullable=False)
-    last_name = Column(String, nullable=False)
-    grade = Column(String)
-    advisor = Column(String)
-    house = Column(String)
-    clan = Column(String)
-
-
-class Teacher(Base):
-    __tablename__ = 'teachers'
-
-    id = Column(String, primary_key=True, default=lambda: str(uuid.uuid4()))
-    name = Column(String, nullable=False)
-    display_name = Column(String)
-
-
-Base.metadata.create_all(bind=engine)
-
-# Persistent storage keys used in the key-value store
-SESSIONS_FILE = 'sessions'
-USERS_FILE = 'users'
-DELETE_REQUESTS_FILE = 'delete_requests'
-INVITE_CODES_FILE = 'invite_codes'
-GLOBAL_CSV_FILE = 'global_csv_data'
-TEACHER_LIST_FILE = 'teacher_list'
-
-
-def load_data_from_file(store_key, default_data):
-    """Load JSON data from the database, falling back to a default."""
-    try:
-        entry = db_session.query(KeyValueStore).filter_by(key=store_key).first()
-        if entry:
-            try:
-                data = json.loads(entry.value)
-                if (isinstance(data, (dict, list)) and not data) and default_data:
-                    return default_data
-                return data
-            except json.JSONDecodeError:
-                print(f"Corrupted data for key {store_key}, resetting to default")
-        serialized = json.dumps(default_data, indent=2)
-        if entry:
-            entry.value = serialized
-        else:
-            db_session.add(KeyValueStore(key=store_key, value=serialized))
-        db_session.commit()
-    except Exception as exc:
-        db_session.rollback()
-        print(f"Error loading {store_key}: {exc}")
-        return default_data
-    return default_data
-
-
-def save_data_to_file(store_key, data):
-    """Persist JSON data to the database-backed key-value store."""
-    try:
-        serialized = json.dumps(data, indent=2)
-        entry = db_session.query(KeyValueStore).filter_by(key=store_key).first()
-        if entry:
-            entry.value = serialized
-        else:
-            db_session.add(KeyValueStore(key=store_key, value=serialized))
-        db_session.commit()
-        print(f"Successfully saved data to {store_key}")
-        return True
-    except Exception as exc:
-        db_session.rollback()
-        print(f"Error saving {store_key}: {exc}")
-        return False
-
-
-DEFAULT_SUPERADMIN = {
-    'username': 'antineutrino',
-    'password': 'b-decay',
-    'role': 'superadmin',
-    'display_name': 'Lead Admin',
-    'status': 'active',
-}
-
-
-def _serialize_user_model(user):
-    if not user:
-        return None
-    return {
-        'id': user.id,
-        'username': user.username,
-        'name': user.display_name,
-        'role': user.role,
-        'status': user.status,
-        'last_login_at': user.last_login_at.isoformat() if user.last_login_at else None,
-    }
-
-
-def get_user_by_username(username):
-    if not username:
-        return None
-    return db_session.query(User).filter(User.username == username).first()
-
-
-def list_all_users():
-    users = db_session.query(User).order_by(User.username.asc()).all()
-    return [_serialize_user_model(user) for user in users]
-
-
-def create_user_record(username, password, display_name, role='user', status='active'):
-    existing = get_user_by_username(username)
-    if existing:
-        return existing
-    user = User(
-        id=str(uuid.uuid4()),
-        username=username,
-        password_hash=password,
-        display_name=display_name,
-        role=role,
-        status=status,
-        created_at=_now_utc(),
-        updated_at=_now_utc(),
-    )
-    try:
-        db_session.add(user)
-        db_session.commit()
-    except Exception:
-        db_session.rollback()
-        raise
-    return user
-
-
-def update_user_credentials(user, *, password=None, display_name=None, role=None, status=None):
-    updated = False
-    if password is not None and user.password_hash != password:
-        user.password_hash = password
-        updated = True
-    if display_name is not None and user.display_name != display_name:
-        user.display_name = display_name
-        updated = True
-    if role is not None and user.role != role:
-        user.role = role
-        updated = True
-    if status is not None and user.status != status:
-        user.status = status
-        updated = True
-    if updated:
-        user.updated_at = _now_utc()
-        try:
-            db_session.commit()
-        except Exception:
-            db_session.rollback()
-            raise
-    return user
-
-
-def ensure_default_superadmin():
-    user = get_user_by_username(DEFAULT_SUPERADMIN['username'])
-    if not user:
-        return create_user_record(
-            DEFAULT_SUPERADMIN['username'],
-            DEFAULT_SUPERADMIN['password'],
-            DEFAULT_SUPERADMIN['display_name'],
-            role=DEFAULT_SUPERADMIN['role'],
-            status=DEFAULT_SUPERADMIN['status'],
-        )
-    return update_user_credentials(
-        user,
-        password=DEFAULT_SUPERADMIN['password'],
-        display_name=DEFAULT_SUPERADMIN['display_name'],
-        role=DEFAULT_SUPERADMIN['role'],
-        status=DEFAULT_SUPERADMIN['status'],
-    )
-
-
-def migrate_legacy_users(legacy_users):
-    if not isinstance(legacy_users, dict):
-        return
-    for username, payload in legacy_users.items():
-        if not isinstance(payload, dict):
-            continue
-        print(f"Migrating legacy user: {username}")
-        password = payload.get('password') or DEFAULT_SUPERADMIN['password']
-        role = payload.get('role', 'user')
-        name = payload.get('name') or username
-        status = payload.get('status', 'active')
-        user = get_user_by_username(username)
-        if user:
-            update_user_credentials(
-                user,
-                password=password,
-                display_name=name,
-                role=role,
-                status=status,
-            )
-        else:
-            create_user_record(
-                username,
-                password,
-                name,
-                role=role,
-                status=status,
-            )
-
-
-def create_invite_code_record(owner_user, issued_by_user, role='user'):
-    code = str(uuid.uuid4())
-    owner = owner_user or issued_by_user
-    invite = UserInviteCode(
-        id=str(uuid.uuid4()),
-        user_id=owner.id,
-        code=code,
-        issued_by=issued_by_user.id,
-        issued_at=_now_utc(),
-        status='unused',
-        role=role,
-    )
-    try:
-        db_session.add(invite)
-        db_session.commit()
-    except Exception:
-        db_session.rollback()
-        raise
-    return invite
-
-
-def get_invite_code_record(code):
-    if not code:
-        return None
-    return db_session.query(UserInviteCode).filter(UserInviteCode.code == code).first()
-
-
-def mark_invite_code_used(invite, used_by_user):
-    invite.status = 'used'
-    invite.user_id = used_by_user.id
-    invite.used_by = used_by_user.id
-    invite.used_at = _now_utc()
-    invite.updated_at = _now_utc()
-    try:
-        db_session.commit()
-    except Exception:
-        db_session.rollback()
-        raise
-
-
-def migrate_legacy_invite_codes(legacy_invites, default_owner):
-    if not isinstance(legacy_invites, dict):
-        return
-    for code, payload in legacy_invites.items():
-        if get_invite_code_record(code):
-            continue
-        role = 'user'
-        status = 'unused'
-        issued_by_username = DEFAULT_SUPERADMIN['username']
-        used_by_username = None
-        if isinstance(payload, dict):
-            role = payload.get('role', 'user')
-            status = 'used' if payload.get('used') else payload.get('status', 'unused')
-            issued_by_username = payload.get('issued_by', DEFAULT_SUPERADMIN['username'])
-            used_by_username = payload.get('used_by')
-        issued_by_user = get_user_by_username(issued_by_username) or default_owner
-        invite = UserInviteCode(
-            id=str(uuid.uuid4()),
-            user_id=default_owner.id,
-            code=code,
-            issued_by=issued_by_user.id,
-            issued_at=_now_utc(),
-            status=status,
-            role=role,
-        )
-        if status == 'used' and used_by_username:
-            used_user = get_user_by_username(used_by_username)
-            if used_user:
-                invite.used_by = used_user.id
-                invite.used_at = _now_utc()
-        try:
-            db_session.add(invite)
-            db_session.commit()
-        except Exception:
-            db_session.rollback()
-            raise
-
-
-def reset_user_store():
-    default_user = ensure_default_superadmin()
-    try:
-        db_session.query(UserInviteCode).delete()
-        db_session.query(User).filter(User.username != default_user.username).delete()
-        db_session.commit()
-    except Exception:
-        db_session.rollback()
-        raise
-    update_user_credentials(
-        default_user,
-        password=DEFAULT_SUPERADMIN['password'],
-        display_name=DEFAULT_SUPERADMIN['display_name'],
-        role=DEFAULT_SUPERADMIN['role'],
-        status=DEFAULT_SUPERADMIN['status'],
-    )
-
-
-# Student lookup cache built from the global CSV data for fast eligibility checks
-student_lookup = {}
-
-secure_random = random.SystemRandom()
-
-
-def normalize_name(value):
-    return str(value or '').strip()
-
-
-def sync_students_table_from_csv_rows(rows):
-    """Persist uploaded student roster into the students table."""
-    result = {'processed': 0, 'created': 0, 'updated': 0}
-    if not isinstance(rows, list) or not rows:
-        return result
-
-    unique_rows = {}
-    for row in rows:
-        if not isinstance(row, dict):
-            continue
-        identifier = normalize_name(row.get('Student ID'))
-        if not identifier:
-            continue
-        key = identifier.lower()
-        if key not in unique_rows:
-            unique_rows[key] = (identifier, row)
-
-    if not unique_rows:
-        return result
-
-    identifiers = [item[0] for item in unique_rows.values()]
-    existing_students = db_session.query(Student).filter(Student.student_identifier.in_(identifiers)).all()
-    existing_map = {student.student_identifier.lower(): student for student in existing_students}
-
-    for key, (identifier, row) in unique_rows.items():
-        preferred = normalize_name(row.get('Preferred'))
-        last = normalize_name(row.get('Last'))
-        if not preferred or not last:
-            continue
-
-        grade = normalize_name(row.get('Grade'))
-        advisor = normalize_name(row.get('Advisor'))
-        house = normalize_name(row.get('House'))
-        clan = normalize_name(row.get('Clan'))
-
-        student = existing_map.get(key)
-        if student:
-            changes = False
-            if student.preferred_name != preferred:
-                student.preferred_name = preferred
-                changes = True
-            if student.last_name != last:
-                student.last_name = last
-                changes = True
-            new_grade = grade or None
-            if student.grade != new_grade:
-                student.grade = new_grade
-                changes = True
-            new_advisor = advisor or None
-            if student.advisor != new_advisor:
-                student.advisor = new_advisor
-                changes = True
-            new_house = house or None
-            if student.house != new_house:
-                student.house = new_house
-                changes = True
-            new_clan = clan or None
-            if student.clan != new_clan:
-                student.clan = new_clan
-                changes = True
-            if changes:
-                result['updated'] += 1
-        else:
-            student = Student(
-                id=str(uuid.uuid4()),
-                student_identifier=identifier,
-                preferred_name=preferred,
-                last_name=last,
-                grade=grade or None,
-                advisor=advisor or None,
-                house=house or None,
-                clan=clan or None,
-            )
-            db_session.add(student)
-            existing_map[key] = student
-            result['created'] += 1
-
-        result['processed'] += 1
-
-    if result['processed']:
-        try:
-            db_session.commit()
-        except Exception:
-            db_session.rollback()
-            raise
-
-    return result
-
-
-def sync_teacher_table_from_list(teachers):
-    """Persist uploaded teacher roster into the teachers table."""
-    result = {'processed': 0, 'created': 0, 'updated': 0}
-    if not isinstance(teachers, list) or not teachers:
-        return result
-
-    unique_entries = {}
-    for entry in teachers:
-        if isinstance(entry, dict):
-            name = normalize_name(entry.get('name'))
-            display = normalize_name(entry.get('display_name') or entry.get('name'))
-        else:
-            name = normalize_name(entry)
-            display = name
-        if not name:
-            continue
-        key = name.lower()
-        if key not in unique_entries:
-            unique_entries[key] = (name, display)
-
-    if not unique_entries:
-        return result
-
-    names = [item[0] for item in unique_entries.values()]
-    existing_teachers = db_session.query(Teacher).filter(Teacher.name.in_(names)).all()
-    existing_map = {teacher.name.lower(): teacher for teacher in existing_teachers}
-
-    for key, (name, display) in unique_entries.items():
-        teacher = existing_map.get(key)
-        desired_display = display or None
-
-        if teacher:
-            if teacher.display_name != desired_display:
-                teacher.display_name = desired_display
-                result['updated'] += 1
-        else:
-            teacher = Teacher(
-                id=str(uuid.uuid4()),
-                name=name,
-                display_name=desired_display,
-            )
-            db_session.add(teacher)
-            existing_map[key] = teacher
-            result['created'] += 1
-
-        result['processed'] += 1
-
-    if result['processed']:
-        try:
-            db_session.commit()
-        except Exception:
-            db_session.rollback()
-            raise
-
-    return result
-
-
-def make_student_key(preferred_name, last_name, student_id=None):
-    student_id_norm = normalize_name(student_id).lower()
-    if student_id_norm:
-        return f"id:{student_id_norm}"
-    preferred_norm = normalize_name(preferred_name).lower()
-    last_norm = normalize_name(last_name).lower()
-    if not preferred_norm and not last_norm:
-        return None
-    return f"{preferred_norm}|{last_norm}"
-
-
-def split_student_key(key):
-    if not key:
-        return '', ''
-    if key.startswith('id:'):
-        return '', ''
-    parts = key.split('|', 1)
-    if len(parts) == 2:
-        return parts[0], parts[1]
-    return parts[0], ''
-
-
-def extract_student_id_from_key(key):
-    if not key:
-        return ''
-    if key.startswith('id:'):
-        return key.split(':', 1)[1]
-    return ''
-
-
-def update_student_lookup():
-    global student_lookup
-    student_lookup = {}
-    data = []
-    if isinstance(global_csv_data, dict):
-        data = global_csv_data.get('data') or []
-    if not isinstance(data, list):
-        data = []
-    for row in data:
-        preferred = normalize_name(row.get('Preferred', ''))
-        last = normalize_name(row.get('Last', ''))
-        student_id = normalize_name(row.get('Student ID', ''))
-        key = make_student_key(preferred, last, student_id)
-        if not key:
-            continue
-        student_lookup[key] = {
-            'preferred_name': preferred,
-            'last_name': last,
-            'grade': normalize_name(row.get('Grade', '')),
-            'advisor': normalize_name(row.get('Advisor', '')),
-            'house': normalize_name(row.get('House', '')),
-            'clan': normalize_name(row.get('Clan', '')),
-            'student_id': student_id,
-            'key': key
-        }
-
-
-def save_all_data():
-    """Save all data to the database-backed store."""
-    try:
-        save_data_to_file(SESSIONS_FILE, session_data)
-        save_data_to_file(DELETE_REQUESTS_FILE, delete_requests)
-        save_data_to_file(GLOBAL_CSV_FILE, global_csv_data)
-        print("All data saved successfully")
-        return True
-    except Exception as e:
-        print(f"Error saving all data: {e}")
-        return False
-
-
-def save_session_data():
-    """Save session data to the database-backed store."""
-    return save_data_to_file(SESSIONS_FILE, session_data)
-
-
-def save_delete_requests():
-    """Save delete requests to the database-backed store."""
-    return save_data_to_file(DELETE_REQUESTS_FILE, delete_requests)
-
-
-def save_global_csv_data():
-    """Save global CSV data to the database-backed store."""
-    result = save_data_to_file(GLOBAL_CSV_FILE, global_csv_data)
-    if result:
-        update_student_lookup()
-    return result
-
-
-def save_global_teacher_data():
-    """Save global teacher data to the database-backed store."""
-    return save_data_to_file(TEACHER_LIST_FILE, global_teacher_data)
-
-
-print("Initializing persistent storage (database-backed)...")
-session_data = load_data_from_file(SESSIONS_FILE, {})
-delete_requests = load_data_from_file(DELETE_REQUESTS_FILE, [])
-legacy_invites = load_data_from_file(INVITE_CODES_FILE, {})
-
-global_csv_data = load_data_from_file(GLOBAL_CSV_FILE, {})
-update_student_lookup()
-
-global_teacher_data = load_data_from_file(TEACHER_LIST_FILE, {})
-
-legacy_users = load_data_from_file(USERS_FILE, {
-    DEFAULT_SUPERADMIN['username']: {
-        'password': DEFAULT_SUPERADMIN['password'],
-        'role': DEFAULT_SUPERADMIN['role'],
-        'name': DEFAULT_SUPERADMIN['display_name'],
-        'status': DEFAULT_SUPERADMIN['status'],
-    }
-})
-
-default_user = ensure_default_superadmin()
-migrate_legacy_users(legacy_users)
-migrate_legacy_invite_codes(legacy_invites, default_user)
-
-print("Saving initial data to ensure tables are seeded...")
-save_data_to_file(SESSIONS_FILE, session_data)
-save_data_to_file(DELETE_REQUESTS_FILE, delete_requests)
-save_data_to_file(GLOBAL_CSV_FILE, global_csv_data)
-save_data_to_file(TEACHER_LIST_FILE, global_teacher_data)
-
-print(f"Initialization complete. Session count: {len(session_data)}, Users: {len(list_all_users())}")
-
-
-def reset_storage_for_testing():
-    """Reset all persistent stores to defaults to keep pytest runs isolated."""
-    global session_data, delete_requests, global_csv_data, global_teacher_data, student_lookup
-
-    session_data = {}
-    delete_requests = []
-    global_csv_data = {}
-    global_teacher_data = {}
-    student_lookup = {}
-
-    save_data_to_file(SESSIONS_FILE, session_data)
-    save_data_to_file(DELETE_REQUESTS_FILE, delete_requests)
-    save_data_to_file(GLOBAL_CSV_FILE, global_csv_data)
-    save_data_to_file(TEACHER_LIST_FILE, global_teacher_data)
-    reset_user_store()
-    update_student_lookup()
-
-
-_last_pytest_identifier = None
-
-
-@recorder_bp.before_app_request
-def _ensure_isolated_state_for_tests():
-    """Clear persisted state when pytest moves to a new test case."""
-    global _last_pytest_identifier
-
-    pytest_identifier = os.environ.get('PYTEST_CURRENT_TEST')
-    if not pytest_identifier:
-        return
-
-    current_identifier = pytest_identifier.split(' (')[0]
-    if current_identifier != _last_pytest_identifier:
-        reset_storage_for_testing()
-        _last_pytest_identifier = current_identifier
-
-
-def ensure_session_structure(session_info):
-    """Ensure session data has the expected structure for counters and records."""
-    if 'dirty_count' not in session_info:
-        dirty_records = session_info.get('dirty_records', [])
-        if isinstance(dirty_records, list):
-            session_info['dirty_count'] = len(dirty_records)
-        else:
-            session_info['dirty_count'] = 0
-
-    # Remove legacy dirty record tracking that stored names
-    if 'dirty_records' in session_info:
-        session_info['dirty_records'] = []
-
-    if 'faculty_clean_records' not in session_info or not isinstance(session_info['faculty_clean_records'], list):
-        session_info['faculty_clean_records'] = []
-
-    if 'scan_history' not in session_info or not isinstance(session_info['scan_history'], list):
-        session_info['scan_history'] = []
-
-    # Strip any stored names from historic dirty scan history entries
-    for record in session_info['scan_history']:
-        if record.get('category') == 'dirty':
-            record.pop('preferred_name', None)
-            record.pop('first_name', None)
-            record.pop('last_name', None)
-            if not record.get('display_name'):
-                record['display_name'] = 'Dirty Plate'
-
-    for category in ['clean_records', 'red_records']:
-        records = session_info.get(category)
-        if not isinstance(records, list):
-            session_info[category] = []
-            continue
-        for record in records:
-            if not isinstance(record, dict):
-                continue
-            preferred = record.get('preferred_name') or record.get('first_name')
-            last = record.get('last_name')
-            student_id = record.get('student_id')
-            key = record.get('student_key')
-            if not key:
-                key = make_student_key(preferred, last, student_id)
-            if key:
-                key_lower = key.lower()
-                record['student_key'] = key_lower
-                if not record.get('student_id'):
-                    extracted = extract_student_id_from_key(key_lower)
-                    if extracted:
-                        record['student_id'] = extracted
-
-    # Ensure draw information structure exists
-    draw_info = session_info.get('draw_info')
-    if not isinstance(draw_info, dict):
-        draw_info = {}
-        session_info['draw_info'] = draw_info
-    draw_info.setdefault('winner', None)
-    draw_info.setdefault('winner_timestamp', None)
-    draw_info.setdefault('selected_by', None)
-    draw_info.setdefault('method', None)
-    draw_info.setdefault('finalized', False)
-    draw_info.setdefault('finalized_at', None)
-    draw_info.setdefault('finalized_by', None)
-    draw_info.setdefault('history', [])
-    draw_info.setdefault('override', False)
-    draw_info.setdefault('tickets_at_selection', None)
-    draw_info.setdefault('probability_at_selection', None)
-    draw_info.setdefault('eligible_pool_size', None)
-    session_info['draw_info'] = draw_info
-
-    if 'is_discarded' not in session_info:
-        session_info['is_discarded'] = False
-    if 'discard_metadata' not in session_info or not isinstance(session_info['discard_metadata'], dict):
-        session_info['discard_metadata'] = {}
-
-
-def get_dirty_count(session_info):
-    """Helper to safely retrieve dirty count from a session."""
-    ensure_session_structure(session_info)
-    return session_info.get('dirty_count', 0)
-
-
-# Normalize any existing sessions loaded from disk
-def normalize_loaded_sessions():
-    for _session_info in session_data.values():
-        ensure_session_structure(_session_info)
-
-
-normalize_loaded_sessions()
-
-
-def format_display_name(profile):
-    preferred = normalize_name(profile.get('preferred_name') or profile.get('first_name', ''))
-    last = normalize_name(profile.get('last_name'))
-    if preferred and last:
-        return f"{preferred} {last}"
-    return preferred or last
-
-
-def build_profile_from_record(record):
-    preferred = normalize_name(record.get('preferred_name') or record.get('first_name'))
-    last = normalize_name(record.get('last_name'))
-    student_id = normalize_name(record.get('student_id'))
-    key = record.get('student_key') or make_student_key(preferred, last, student_id)
-    if not student_id and key:
-        student_id = extract_student_id_from_key(key)
-    profile = {
-        'preferred_name': preferred,
-        'last_name': last,
-        'grade': normalize_name(record.get('grade')),
-        'advisor': normalize_name(record.get('advisor')),
-        'house': normalize_name(record.get('house')),
-        'clan': normalize_name(record.get('clan')),
-        'student_id': student_id
-    }
-    profile['key'] = key
-    if key and key in student_lookup:
-        lookup = student_lookup[key]
-        for field in ['preferred_name', 'last_name', 'grade', 'advisor', 'house', 'clan', 'student_id']:
-            if not profile.get(field):
-                profile[field] = lookup.get(field, '')
-    profile['display_name'] = format_display_name(profile)
-    return profile
-
-
-def is_student_profile_eligible(profile):
-    key = profile.get('key')
-    if not key:
-        return False
-    if student_lookup:
-        return key in student_lookup
-    return True
-
-
-def safe_parse_datetime(value):
-    if not value:
-        return datetime.min
-    try:
-        return datetime.fromisoformat(value)
-    except ValueError:
-        try:
-            return datetime.fromisoformat(value.replace('Z', '+00:00'))
-        except Exception:
-            return datetime.min
-
-
-def compute_ticket_rollups():
-    generated_at = datetime.now().isoformat()
-    ordered = sorted(
-        session_data.items(),
-        key=lambda item: (safe_parse_datetime(item[1].get('created_at')), item[0])
-    )
-    summaries = {}
-    current_tickets = {}
-    student_profiles = {}
-    for session_id, info in ordered:
-        ensure_session_structure(info)
-        if info.get('is_discarded'):
-            summaries[session_id] = {
-                'session_id': session_id,
-                'session_name': info.get('session_name', ''),
-                'created_at': info.get('created_at'),
-                'is_discarded': True,
-                'tickets_snapshot': {},
-                'profiles': {},
-                'total_tickets': 0.0,
-                'candidates': [],
-                'top_candidates': [],
-                'eligible_count': 0,
-                'excluded_records': 0,
-                'generated_at': generated_at
-            }
-            continue
-        pre_session_keys = set(current_tickets.keys())
-        present_keys = set()
-        excluded_records = 0
-        for record in info.get('clean_records', []):
-            profile = build_profile_from_record(record)
-            key = profile.get('key')
-            if not key or not is_student_profile_eligible(profile):
-                excluded_records += 1
-                continue
-            present_keys.add(key)
-            student_profiles[key] = profile
-            current_tickets[key] = current_tickets.get(key, 0.0) + 1.0
-        for record in info.get('red_records', []):
-            profile = build_profile_from_record(record)
-            key = profile.get('key')
-            if not key or not is_student_profile_eligible(profile):
-                excluded_records += 1
-                continue
-            present_keys.add(key)
-            student_profiles[key] = profile
-            current_tickets[key] = 0.0
-        for key in pre_session_keys:
-            if key not in present_keys:
-                value = current_tickets.get(key, 0.0)
-                if value > 0:
-                    current_tickets[key] = value / 2.0
-        snapshot = {k: float(v) for k, v in current_tickets.items() if v > 0}
-        total_tickets = sum(snapshot.values())
-        profiles_snapshot = {k: student_profiles.get(k, {}).copy() for k in snapshot}
-        candidates = []
-        for key, value in sorted(snapshot.items(), key=lambda item: (-item[1], item[0])):
-            profile = profiles_snapshot.get(key, {})
-            display_name = profile.get('display_name') or format_display_name(profile)
-            candidate = {
-                'key': key,
-                'tickets': value,
-                'preferred_name': profile.get('preferred_name', ''),
-                'last_name': profile.get('last_name', ''),
-                'display_name': display_name,
-                'grade': profile.get('grade', ''),
-                'advisor': profile.get('advisor', ''),
-                'house': profile.get('house', ''),
-                'clan': profile.get('clan', ''),
-                'student_id': profile.get('student_id', ''),
-                'probability': (value / total_tickets * 100.0) if total_tickets > 0 else 0.0
-            }
-            candidates.append(candidate)
-        summaries[session_id] = {
-            'session_id': session_id,
-            'session_name': info.get('session_name', ''),
-            'created_at': info.get('created_at'),
-            'is_discarded': False,
-            'tickets_snapshot': snapshot,
-            'profiles': profiles_snapshot,
-            'total_tickets': total_tickets,
-            'candidates': candidates,
-            'top_candidates': candidates[:3],
-            'eligible_count': len(candidates),
-            'excluded_records': excluded_records,
-            'generated_at': generated_at
-        }
-        draw_info = info.get('draw_info', {})
-        winner = draw_info.get('winner')
-        if winner and draw_info.get('finalized'):
-            winner_key = winner.get('key')
-            if winner_key:
-                current_tickets[winner_key] = 0.0
-    return summaries
-
-
-def get_ticket_summary_for_session(session_id):
-    summaries = compute_ticket_rollups()
-    return summaries.get(session_id), summaries
-
-
-def serialize_draw_info(draw_info):
-    if not isinstance(draw_info, dict):
-        return {
-            'winner': None,
-            'winner_timestamp': None,
-            'selected_by': None,
-            'method': None,
-            'finalized': False,
-            'finalized_at': None,
-            'finalized_by': None,
-            'override': False,
-            'tickets_at_selection': None,
-            'probability_at_selection': None,
-            'eligible_pool_size': None,
-            'history': []
-        }
-    result = {
-        'winner_timestamp': draw_info.get('winner_timestamp'),
-        'selected_by': draw_info.get('selected_by'),
-        'method': draw_info.get('method'),
-        'finalized': draw_info.get('finalized', False),
-        'finalized_at': draw_info.get('finalized_at'),
-        'finalized_by': draw_info.get('finalized_by'),
-        'override': draw_info.get('override', False),
-        'tickets_at_selection': draw_info.get('tickets_at_selection'),
-        'probability_at_selection': draw_info.get('probability_at_selection'),
-        'eligible_pool_size': draw_info.get('eligible_pool_size'),
-        'history': list(draw_info.get('history', []))
-    }
-    winner = draw_info.get('winner')
-    if isinstance(winner, dict):
-        result['winner'] = {
-            'key': winner.get('key'),
-            'display_name': winner.get('display_name') or format_display_name(winner),
-            'preferred_name': winner.get('preferred_name'),
-            'last_name': winner.get('last_name'),
-            'grade': winner.get('grade'),
-            'advisor': winner.get('advisor'),
-            'house': winner.get('house'),
-            'clan': winner.get('clan'),
-            'student_id': winner.get('student_id'),
-            'tickets': winner.get('tickets'),
-            'probability': winner.get('probability')
-        }
-    else:
-        result['winner'] = None
-    return result
-
-
-def get_current_user():
-    """Get current logged in user"""
-    username = session.get('user_id')
-    if not username:
-        return None
-    user = get_user_by_username(username)
-    return _serialize_user_model(user)
-
-
-def require_auth():
-    """Check if user is authenticated"""
-    username = session.get('user_id')
-    if not username:
-        return False
-    return get_user_by_username(username) is not None
-
-
-def require_auth_or_guest():
-    """Check if user is authenticated or is a guest"""
-    return require_auth() or session.get('guest_access', False)
-
-
-def is_guest():
-    """Check if current user is a guest"""
-    return session.get('guest_access', False) and 'user_id' not in session
-
-
-def require_admin():
-    """Check if user is admin or super admin"""
-    user = get_current_user()
-    return user and user['role'] in ['admin', 'superadmin']
-
-
-def require_superadmin():
-    """Check if user is super admin"""
-    user = get_current_user()
-    return user and user['role'] == 'superadmin'
-
-
-@recorder_bp.route('/auth/login', methods=['POST'])
-def login():
-    """User login"""
-    data = request.get_json()
-    username = data.get('username', '').strip()
-    password = data.get('password', '').strip()
-
-    user = get_user_by_username(username)
-    if not user or user.password_hash != password:
-        return jsonify({'error': 'Invalid username or password'}), 401
-
-    if user.status != 'active':
-        return jsonify({'error': 'Account is disabled. Please contact an administrator.'}), 403
-
-    session['user_id'] = username
-    user.last_login_at = _now_utc()
-    user.updated_at = _now_utc()
-    try:
-        db_session.commit()
-    except Exception:
-        db_session.rollback()
-        return jsonify({'error': 'Unable to update login information'}), 500
-
-    return jsonify({
-        'status': 'success',
-        'user': {
-            'username': username,
-            'name': user.display_name,
-            'role': user.role
-        }
-    }), 200
-
-
-@recorder_bp.route('/auth/logout', methods=['POST'])
-def logout():
-    """User logout"""
-    user_id = session.pop('user_id', None)
-    session.pop('session_id', None)
-    session.pop('guest_access', None)
-    return jsonify({'status': 'success', 'message': 'Logged out successfully'}), 200
-
-
-@recorder_bp.route('/auth/guest', methods=['POST'])
-def guest_login():
-    """Guest login - allows viewing sessions without signup"""
-    session['guest_access'] = True
-    return jsonify({
-        'status': 'success',
-        'user': {
-            'username': 'guest',
-            'name': 'Guest User',
-            'role': 'guest'
-        }
-    }), 200
-
-
-@recorder_bp.route('/auth/signup', methods=['POST'])
-def signup():
-    """User signup"""
-    data = request.get_json()
-    username = data.get('username', '').strip()
-    password = data.get('password', '').strip()
-    name = data.get('name', '').strip()
-    invite_code = data.get('invite_code', '').strip()
-
-    # Validation
-    if not username or not password or not name or not invite_code:
-        return jsonify({'error': 'Username, password, name, and invite code are required'}), 400
-
-    if len(username) < 3:
-        return jsonify({'error': 'Username must be at least 3 characters long'}), 400
-
-    if len(password) < 6:
-        return jsonify({'error': 'Password must be at least 6 characters long'}), 400
-
-    if get_user_by_username(username):
-        return jsonify({'error': 'Username already exists'}), 409
-
-    # Validate invite code
-    invite = get_invite_code_record(invite_code)
-    if not invite or invite.status != 'unused':
-        return jsonify({'error': 'Invalid invite code'}), 403
-
-    if invite.expires_at and invite.expires_at < _now_utc():
-        return jsonify({'error': 'Invite code has expired'}), 403
-
-    try:
-        new_user = create_user_record(
-            username,
-            password,
-            name,
-            role=invite.role or 'user',
-            status='active'
-        )
-    except Exception:
-        return jsonify({'error': 'Could not create user'}), 500
-
-    try:
-        mark_invite_code_used(invite, new_user)
-    except Exception:
-        # Roll back user creation if invite update fails
-        user_to_delete = get_user_by_username(username)
-        if user_to_delete:
-            try:
-                db_session.delete(user_to_delete)
-                db_session.commit()
-            except Exception:
-                db_session.rollback()
-        return jsonify({'error': 'Could not update invite code'}), 500
-
-    return jsonify({
-        'status': 'success',
-        'message': 'Account created successfully'
-    }), 201
-
-
-@recorder_bp.route('/auth/status', methods=['GET'])
-def auth_status():
-    """Check authentication status"""
-    if require_auth():
-        user = get_current_user()
-        return jsonify({
-            'authenticated': True,
-            'user': {
-                'username': session['user_id'],
-                'name': user['name'],
-                'role': user['role']
-            }
-        }), 200
-    elif is_guest():
-        return jsonify({
-            'authenticated': True,
-            'user': {
-                'username': 'guest',
-                'name': 'Guest User',
-                'role': 'guest'
-            }
-        }), 200
-    else:
-        return jsonify({'authenticated': False}), 200
-
-
-@recorder_bp.route('/admin/manage-account-status', methods=['POST'])
-def manage_account_status():
-    """Manage account status (enable/disable users)"""
-    if not require_auth():
-        return jsonify({'error': 'Authentication required'}), 401
-
-    current_user = get_current_user()
-    data = request.get_json()
-    target_username = data.get('username')
-    new_status = data.get('status')  # 'active' or 'disabled'
-
-    if not target_username or not new_status:
-        return jsonify({'error': 'Username and status are required'}), 400
-
-    target_user_model = get_user_by_username(target_username)
-    if not target_user_model:
-        return jsonify({'error': 'User not found'}), 404
-
-    target_user = _serialize_user_model(target_user_model)
-
-    # Permission checks
-    if current_user['role'] == 'superadmin':
-        # Super admin can manage everyone except themselves
-        if target_username == session['user_id']:
-            return jsonify({'error': 'Cannot modify your own account status'}), 403
-    elif current_user['role'] == 'admin':
-        # Admin can manage users but not super admins or other admins
-        if target_user['role'] in ['superadmin', 'admin']:
-            return jsonify({'error': 'Insufficient permissions to modify this account'}), 403
-    else:
-        return jsonify({'error': 'Insufficient permissions'}), 403
-
-    # Update status
-    try:
-        update_user_credentials(target_user_model, status=new_status)
-    except Exception:
-        return jsonify({'error': 'Failed to update user status'}), 500
-
-    return jsonify({
-        'status': 'success',
-        'message': f'Account {target_username} has been {new_status}'
-    }), 200
-
-
-@recorder_bp.route('/session/request-delete', methods=['POST'])
-def request_delete_session():
-    """Submit a delete request for a session"""
-    if not require_auth():
-        return jsonify({'error': 'Authentication required'}), 401
-
-    data = request.get_json()
-    session_id = data.get('session_id')
-
-    if not session_id:
-        return jsonify({'error': 'Session ID is required'}), 400
-
-    if session_id not in session_data:
-        return jsonify({'error': 'Session not found'}), 404
-
-    # Get current user
-    current_user = get_current_user()
-
-    # Users can request deletion for any session; admins and superadmins delete immediately
-
-    # Admins and super admins delete immediately
-    if current_user['role'] in ['admin', 'superadmin']:
-        session_name = session_data[session_id]['session_name']
-        del session_data[session_id]
-        save_session_data()
-        if session.get('session_id') == session_id:
-            session.pop('session_id', None)
-        return jsonify({
-            'status': 'success',
-            'message': f'Session "{session_name}" deleted successfully',
-            'deleted_session_id': session_id
-        }), 200
-
-    # Prevent multiple delete requests for the same session
-    existing = next((req for req in delete_requests
-                     if req['session_id'] == session_id and req['status'] == 'pending'), None)
-    if existing:
-        return jsonify({'error': 'Delete request already submitted for this session'}), 400
-
-    session_info = session_data[session_id]
-    ensure_session_structure(session_info)
-    session_name = session_info['session_name']
-
-    # Collect session statistics
-    clean_count = len(session_info.get('clean_records', []))
-    dirty_count = get_dirty_count(session_info)
-    red_count = len(session_info.get('red_records', []))
-    faculty_clean_count = len(session_info.get('faculty_clean_records', []))
-    total_records = clean_count + dirty_count + red_count
-
-    # Create delete request
-    request_obj = {
-        'id': str(uuid.uuid4()),
-        'session_id': session_id,
-        'session_name': session_name,
-        'requester': session['user_id'],
-        'requester_name': current_user['name'],
-        'requested_at': datetime.now().isoformat(),
-        'status': 'pending',
-        'total_records': total_records,
-        'clean_records': clean_count,
-        'dirty_records': dirty_count,
-        'red_records': red_count,
-        'faculty_clean_records': faculty_clean_count,
-    }
-
-    delete_requests.append(request_obj)
-    save_delete_requests()
-
-    return jsonify({
-        'status': 'success',
-        'message': f'Delete request submitted for "{session_name}"',
-        'request': request_obj
-    }), 200
-
-
-@recorder_bp.route('/admin/delete-requests', methods=['GET'])
-def get_delete_requests():
-    """Get pending delete requests (admin/super admin only)"""
-    if not require_admin():
-        return jsonify({'error': 'Admin access required'}), 403
-
-    pending_requests = [req for req in delete_requests if req.get('status') == 'pending']
-    return jsonify({
-        'status': 'success',
-        'requests': pending_requests
-    }), 200
-
-
-@recorder_bp.route('/admin/users', methods=['GET'])
-def admin_get_users():
-    """Admin: Get all users"""
-    if not require_admin():
-        return jsonify({'error': 'Admin access required'}), 403
-
-    users_list = []
-    for user in list_all_users():
-        users_list.append({
-            'username': user['username'],
-            'name': user['name'],
-            'role': user['role']
-        })
-
-    return jsonify({'users': users_list}), 200
-
-
-@recorder_bp.route('/admin/invite', methods=['POST'])
-def admin_create_invite():
-    """Admin: Generate a one-time invite code"""
-    if not require_admin():
-        return jsonify({'error': 'Admin access required'}), 403
-
-    issuer = get_user_by_username(session.get('user_id'))
-    if not issuer:
-        return jsonify({'error': 'Unable to locate issuing user'}), 500
-
-    invite = create_invite_code_record(issuer, issuer, role='user')
-    return jsonify({'status': 'success', 'invite_code': invite.code}), 201
-
-
-@recorder_bp.route('/admin/sessions', methods=['GET'])
-def admin_get_all_sessions():
-    """Admin: Get all sessions from all users"""
-    if not require_admin():
-        return jsonify({'error': 'Admin access required'}), 403
-
-    all_sessions = []
-    for session_id, data in session_data.items():
-        ensure_session_structure(data)
-        total_records = len(data['clean_records']) + get_dirty_count(data) + len(data['red_records'])
-        all_sessions.append({
-            'session_id': session_id,
-            'session_name': data['session_name'],
-            'owner': data.get('owner', 'unknown'),
-            'created_at': data.get('created_at', 'unknown'),
-            'total_records': total_records,
-            'clean_count': len(data['clean_records']),
-            'dirty_count': get_dirty_count(data),
-            'red_count': len(data['red_records']),
-            'faculty_clean_count': len(data.get('faculty_clean_records', []))
-        })
-
-    return jsonify({'sessions': all_sessions}), 200
-
-
-@recorder_bp.route('/admin/sessions/<session_id>', methods=['DELETE'])
-def admin_delete_session(session_id):
-    """Admin: Delete any session"""
-    if not require_admin():
-        return jsonify({'error': 'Admin access required'}), 403
-
-    if session_id not in session_data:
-        return jsonify({'error': 'Session not found'}), 404
-
-    session_name = session_data[session_id]['session_name']
-    del session_data[session_id]
-    save_session_data()
-    if session.get('session_id') == session_id:
-        session.pop('session_id', None)
-
-    return jsonify({
-        'status': 'success',
-        'message': f'Session "{session_name}" deleted successfully by admin',
-        'deleted_session_id': session_id
-    }), 200
-
-
-@recorder_bp.route('/session/create', methods=['POST'])
-def create_session():
-    """Create a new session"""
-    if not require_auth():
-        return jsonify({'error': 'Authentication required'}), 401
-
-    data = request.get_json(silent=True) or {}
-    custom_name = data.get('session_name', '').strip()
-    is_public = data.get('is_public', True)
-
-    # Generate session ID
-    session_id = str(uuid.uuid4())
-
-    # Existing session names
-    existing_names = {d['session_name'] for d in session_data.values()}
-
-    # Generate session name
-    if custom_name:
-        if custom_name in existing_names:
-            return jsonify({'error': 'Session name already exists'}), 400
-        session_name = custom_name
-    else:
-        now = datetime.now()
-        base_name = f"Golden_Plate_{now.strftime('%B_%d_%Y')}"
-        session_name = base_name
-        counter = 1
-        while session_name in existing_names:
-            session_name = f"{base_name}_{counter}"
-            counter += 1
-
-    # Create session data with owner information
-    session_data[session_id] = {
-        'session_name': session_name,
-        'owner': session['user_id'],
-        'created_at': datetime.now().isoformat(),
-        'clean_records': [],
-        'dirty_count': 0,
-        'red_records': [],
-        'faculty_clean_records': [],
-        'scan_history': [],
-        'is_public': is_public,
-        'draw_info': {
-            'winner': None,
-            'winner_timestamp': None,
-            'selected_by': None,
-            'method': None,
-            'finalized': False,
-            'finalized_at': None,
-            'finalized_by': None,
-            'history': [],
-            'override': False,
-            'tickets_at_selection': None,
-            'probability_at_selection': None,
-            'eligible_pool_size': None
-        },
-        'is_discarded': False,
-        'discard_metadata': {}
-    }
-
-    # Save session data to store
-    save_session_data()
-
-    # Set as current session
-    session['session_id'] = session_id
-
-    return jsonify({
-        'session_id': session_id,
-        'session_name': session_name,
-        'owner': session['user_id']
-    }), 200
-
-
-@recorder_bp.route('/session/list', methods=['GET'])
-def list_sessions():
-    """List sessions. Guests only see public ones."""
-    if not require_auth_or_guest():
-        return jsonify({'error': 'Authentication or guest access required'}), 401
-
-    user_sessions = []
-    for session_id, data in session_data.items():
-        ensure_session_structure(data)
-        if is_guest() and not data.get('is_public', True):
-            continue
-        total_records = len(data['clean_records']) + get_dirty_count(data) + len(data['red_records'])
-        clean_count = len(data['clean_records'])
-        dirty_count = get_dirty_count(data) + len(data['red_records'])  # Combine dirty + very dirty
-        faculty_clean_count = len(data.get('faculty_clean_records', []))
-
-        # Calculate percentages - both clean and faculty clean count as "clean" for ratio
-        combined_clean_count = clean_count + faculty_clean_count
-        total_for_ratio = total_records + faculty_clean_count  # Include faculty in total for ratio calculation
-        clean_percentage = (combined_clean_count / total_for_ratio * 100) if total_for_ratio > 0 else 0
-        dirty_percentage = (dirty_count / total_for_ratio * 100) if total_for_ratio > 0 else 0
-
-        pending = any(req['session_id'] == session_id and req['status'] == 'pending'
-                       for req in delete_requests)
-        user_sessions.append({
-            'session_id': session_id,
-            'session_name': data['session_name'],
-            'owner': data.get('owner', 'unknown'),
-            'total_records': total_records,
-            'clean_count': clean_count,
-            'dirty_count': dirty_count,
-            'faculty_clean_count': faculty_clean_count,
-            'clean_percentage': round(clean_percentage, 1),
-            'dirty_percentage': round(dirty_percentage, 1),
-            'is_public': data.get('is_public', True),
-            'delete_requested': pending,
-            'is_discarded': data.get('is_discarded', False),
-            'draw_info': serialize_draw_info(data.get('draw_info', {}))
-        })
-
-    return jsonify({
-        'sessions': user_sessions,
-        'has_global_csv': bool(global_csv_data.get('data'))
-    }), 200
-
-
-@recorder_bp.route('/session/switch/<session_id>', methods=['POST'])
-def switch_session(session_id):
-    """Switch to a different session. Guests may only access public sessions."""
-    if not require_auth_or_guest():
-        return jsonify({'error': 'Authentication or guest access required'}), 401
-
-    if session_id not in session_data:
-        return jsonify({'error': 'Session not found'}), 404
-
-    if is_guest() and not session_data[session_id].get('is_public', True):
-        return jsonify({'error': 'Access denied'}), 403
-
-    session['session_id'] = session_id
-
-    return jsonify({
-        'session_id': session_id,
-        'session_name': session_data[session_id]['session_name']
-    }), 200
-
-
-@recorder_bp.route('/session/delete/<session_id>', methods=['DELETE'])
-def delete_session(session_id):
-    """Delete a session directly"""
-    if not require_auth():
-        return jsonify({'error': 'Authentication required'}), 401
-
-    if session_id not in session_data:
-        return jsonify({'error': 'Session not found'}), 404
-
-    # Get current user
-    current_user = get_current_user()
-    session_owner = session_data[session_id].get('owner')
-
-    # Permission check: users can only delete their own sessions, admins and super admins can delete any session
-    if current_user['role'] == 'user' and session_owner != session['user_id']:
-        return jsonify({'error': 'You can only delete sessions that you created'}), 403
-
-    session_name = session_data[session_id]['session_name']
-    del session_data[session_id]
-    save_session_data()
-    if session.get('session_id') == session_id:
-        session.pop('session_id', None)
-
-    return jsonify({
-        'status': 'success',
-        'message': f'Session "{session_name}" deleted successfully',
-        'deleted_session_id': session_id
-    }), 200
-
-
-@recorder_bp.route('/admin/delete-requests/<request_id>/approve', methods=['POST'])
-def approve_delete_request(request_id):
-    """Approve a delete request (admin/superadmin only)"""
-    if not require_admin():
-        return jsonify({'error': 'Admin access required'}), 403
-
-    # Find the request
-    request_obj = next((req for req in delete_requests if req['id'] == request_id), None)
-    if not request_obj:
-        return jsonify({'error': 'Delete request not found'}), 404
-
-    if request_obj['status'] != 'pending':
-        return jsonify({'error': 'Request is not pending'}), 400
-
-    session_id = request_obj['session_id']
-
-    # Check if session still exists
-    if session_id not in session_data:
-        # Mark request as completed since session no longer exists
-        request_obj['status'] = 'completed'
-        request_obj['approved_by'] = session['user_id']
-        request_obj['approved_at'] = datetime.now().isoformat()
-        save_delete_requests()
-        return jsonify({'message': 'Session no longer exists, request marked as completed'}), 200
-
-    # Delete the session
-    session_name = session_data[session_id]['session_name']
-    del session_data[session_id]
-    save_session_data()
-    if session.get('session_id') == session_id:
-        session.pop('session_id', None)
-
-    # Update request status
-    request_obj['status'] = 'approved'
-    request_obj['approved_by'] = session['user_id']
-    request_obj['approved_at'] = datetime.now().isoformat()
-    save_delete_requests()
-
-    return jsonify({
-        'status': 'success',
-        'message': f'Session "{session_name}" deleted successfully',
-        'deleted_session_id': session_id
-    }), 200
-
-
-@recorder_bp.route('/admin/approve-delete', methods=['POST'])
-def approve_delete_request_api():
-    data = request.get_json(silent=True) or {}
-    request_id = data.get('request_id')
-    if not request_id:
-        return jsonify({'error': 'Request ID is required'}), 400
-    return approve_delete_request(request_id)
-
-
-@recorder_bp.route('/admin/delete-requests/<request_id>/reject', methods=['POST'])
-def reject_delete_request(request_id):
-    """Reject a delete request (admin/superadmin only)"""
-    if not require_admin():
-        return jsonify({'error': 'Admin access required'}), 403
-
-    data = request.get_json(silent=True) or {}
-    rejection_reason = data.get('reason', 'No reason provided')
-
-    # Find the request
-    request_obj = next((req for req in delete_requests if req['id'] == request_id), None)
-    if not request_obj:
-        return jsonify({'error': 'Delete request not found'}), 404
-
-    if request_obj['status'] != 'pending':
-        return jsonify({'error': 'Request is not pending'}), 400
-
-    # Update request status
-    request_obj['status'] = 'rejected'
-    request_obj['rejected_by'] = session['user_id']
-    request_obj['rejected_at'] = datetime.now().isoformat()
-    request_obj['rejection_reason'] = rejection_reason
-    save_delete_requests()
-
-    return jsonify({
-        'status': 'success',
-        'message': f'Delete request rejected',
-        'request_id': request_id
-    }), 200
-
-
-@recorder_bp.route('/csv/upload', methods=['POST'])
-def upload_csv():
-    """Upload CSV file (requires admin or super admin)"""
-    if not require_admin():
-        return jsonify({'error': 'Admin or super admin access required'}), 403
-
-    if 'file' not in request.files:
-        return jsonify({'error': 'No file provided'}), 400
-
-    file = request.files['file']
-    if file.filename == '':
-        return jsonify({'error': 'No file selected'}), 400
-
-    if not file.filename.lower().endswith('.csv'):
-        return jsonify({'error': 'File must be a CSV'}), 400
-
-    try:
-        # Read CSV content
-        content = file.read().decode('utf-8')
-        csv_reader = csv.DictReader(io.StringIO(content))
-
-        # Convert to list and validate structure
-        rows = list(csv_reader)
-        if not rows:
-            return jsonify({'error': 'CSV file is empty'}), 400
-
-        # Check for required columns
-        required_columns = ['Student ID', 'Last', 'Preferred', 'Grade', 'Advisor', 'House', 'Clan']
-        if not all(col in csv_reader.fieldnames for col in required_columns):
-            return jsonify({'error': f'CSV must contain columns: {', '.join(required_columns)}'}), 400
-
-        # Store globally (accessible to all users)
-        user_id = session['user_id']
-        global global_csv_data
-        global_csv_data = {
-            'data': rows,
-            'columns': csv_reader.fieldnames,
-            'uploaded_by': user_id,
-            'uploaded_at': datetime.now().isoformat()
-        }
-        update_student_lookup()
-
-        # Save to persistent storage
-        save_global_csv_data()
-
-        try:
-            sync_result = sync_students_table_from_csv_rows(rows)
-            print(f"Student roster sync complete: {sync_result}")
-        except Exception as e:
-            print(f"Error syncing students table: {e}")
-            return jsonify({'error': 'Student roster could not be stored in the database'}), 500
-
-        return jsonify({
-            'status': 'success',
-            'rows_count': len(rows),
-            'uploaded_by': user_id,
-            'students_processed': sync_result['processed'],
-            'students_created': sync_result['created'],
-            'students_updated': sync_result['updated']
-        }), 200
-
-    except Exception as e:
-        return jsonify({'error': f'Error processing CSV: {str(e)}'}), 400
-
-
-@recorder_bp.route('/csv/preview', methods=['GET'])
-def preview_csv():
-    """Preview the current student database (admin/super admin only)"""
-    if not require_admin():
-        return jsonify({'error': 'Admin or super admin access required'}), 403
-
-    csv_data = global_csv_data
-
-    if not csv_data:
-        return jsonify({
-            'status': 'no_data',
-            'message': 'No student database uploaded yet'
-        }), 200
-
-    # Get pagination parameters
-    page = int(request.args.get('page', 1))
-    per_page = int(request.args.get('per_page', 50))
-
-    # Calculate pagination
-    total_records = len(csv_data['data'])
-    start_idx = (page - 1) * per_page
-    end_idx = start_idx + per_page
-
-    # Get paginated data
-    paginated_rows = csv_data['data'][start_idx:end_idx]
-    sanitized_rows = [{
-        'Preferred': str(row.get('Preferred', '') or '').strip(),
-        'Last': str(row.get('Last', '') or '').strip(),
-        'Grade': str(row.get('Grade', '') or '').strip(),
-        'Advisor': str(row.get('Advisor', '') or '').strip(),
-        'House': str(row.get('House', '') or '').strip(),
-        'Clan': str(row.get('Clan', '') or '').strip()
-    } for row in paginated_rows]
-
-    return jsonify({
-        'status': 'success',
-        'data': sanitized_rows,
-        'columns': ['Preferred', 'Last', 'Grade', 'Advisor', 'House', 'Clan'],
-        'pagination': {
-            'page': page,
-            'per_page': per_page,
-            'total_records': total_records,
-            'total_pages': (total_records + per_page - 1) // per_page,
-            'has_next': end_idx < total_records,
-            'has_prev': page > 1
-        },
-        'metadata': {
-            'uploaded_by': csv_data.get('uploaded_by', 'unknown'),
-            'uploaded_at': csv_data.get('uploaded_at', 'unknown')
-        }
-    }), 200
-
-
-@recorder_bp.route('/csv/student-names', methods=['GET'])
-def get_student_names():
-    """Get student names for dropdown suggestions"""
-    if not require_auth():
-        return jsonify({'error': 'Authentication required'}), 401
-
-    csv_data = global_csv_data
-    if not csv_data or 'data' not in csv_data:
-        return jsonify({
-            'status': 'no_data',
-            'names': []
-        }), 200
-
-    # Extract names from CSV data
-    names = []
-    for row in csv_data['data']:
-        preferred = str(row.get('Preferred', '') or '').strip()
-        last = str(row.get('Last', '') or '').strip()
-        student_id = str(row.get('Student ID', '') or '').strip()
-        key = make_student_key(preferred, last, student_id)
-
-        if preferred and last:
-            full_name = f"{preferred} {last}"
-            names.append({
-                'display_name': full_name,
-                'preferred': preferred,
-                'preferred_name': preferred,
-                'last': last,
-                'last_name': last,
-                'student_id': student_id,
-                'key': key
-            })
-        elif preferred:
-            names.append({
-                'display_name': preferred,
-                'preferred': preferred,
-                'preferred_name': preferred,
-                'last': '',
-                'last_name': '',
-                'student_id': student_id,
-                'key': key
-            })
-
-    # Sort names alphabetically
-    names.sort(key=lambda x: x['display_name'].lower())
-
-    return jsonify({
-        'status': 'success',
-        'names': names
-    }), 200
-
-
-@recorder_bp.route('/teachers/upload', methods=['POST'])
-def upload_teachers():
-    """Upload teacher list (admin/super admin only)"""
-    if not require_admin():
-        return jsonify({'error': 'Admin or super admin access required'}), 403
-
-    if 'file' not in request.files:
-        return jsonify({'error': 'No file uploaded'}), 400
-
-    file = request.files['file']
-    if file.filename == '':
-        return jsonify({'error': 'No file selected'}), 400
-
-    if not file.filename.lower().endswith(('.csv', '.txt')):
-        return jsonify({'error': 'File must be a CSV or TXT file'}), 400
-
-    try:
-        # Read file content
-        content = file.read().decode('utf-8-sig')  # Handle BOM if present
-        lines = content.strip().split('\n')
-
-        teachers = []
-        for line in lines:
-            # Remove quotes and whitespace
-            teacher_name = line.strip().strip('"').strip("'")
-            if teacher_name:  # Skip empty lines
-                teachers.append({
-                    'name': teacher_name,
-                    'display_name': teacher_name
-                })
-
-        if not teachers:
-            return jsonify({'error': 'No valid teacher names found in file'}), 400
-
-        # Store globally (accessible to all users)
-        user_id = session['user_id']
-        global global_teacher_data
-        global_teacher_data = {
-            'teachers': teachers,
-            'uploaded_by': user_id,
-            'uploaded_at': datetime.now().isoformat()
-        }
-
-        # Save to persistent storage
-        save_global_teacher_data()
-
-        try:
-            teacher_sync = sync_teacher_table_from_list(teachers)
-            print(f"Teacher roster sync complete: {teacher_sync}")
-        except Exception as e:
-            print(f"Error syncing teachers table: {e}")
-            return jsonify({'error': 'Teacher roster could not be stored in the database'}), 500
-
-        return jsonify({
-            'status': 'success',
-            'count': len(teachers),
-            'uploaded_by': user_id,
-            'teachers_processed': teacher_sync['processed'],
-            'teachers_created': teacher_sync['created'],
-            'teachers_updated': teacher_sync['updated']
-        }), 200
-
-    except Exception as e:
-        return jsonify({'error': f'Error processing file: {str(e)}'}), 400
-
-
-@recorder_bp.route('/teachers/list', methods=['GET'])
-def get_teacher_names():
-    """Get teacher names for dropdown suggestions"""
-    if not require_auth():
-        return jsonify({'error': 'Authentication required'}), 401
-
-    teacher_data = global_teacher_data
-    if not teacher_data or 'teachers' not in teacher_data:
-        return jsonify({
-            'status': 'no_data',
-            'names': []
-        }), 200
-
-    # Sort names alphabetically
-    teachers = sorted(teacher_data['teachers'], key=lambda x: x['display_name'].lower())
-
-    return jsonify({
-        'status': 'success',
-        'names': teachers
-    }), 200
-
-
-@recorder_bp.route('/teachers/preview', methods=['GET'])
-def preview_teachers():
-    """Preview the current teacher list (admin/super admin only)"""
-    if not require_admin():
-        return jsonify({'error': 'Admin or super admin access required'}), 403
-
-    teacher_data = global_teacher_data
-
-    if not teacher_data:
-        return jsonify({
-            'status': 'no_data',
-            'message': 'No teacher list uploaded yet'
-        }), 200
-
-    # Get pagination parameters
-    page = int(request.args.get('page', 1))
-    per_page = int(request.args.get('per_page', 50))
-
-    teachers = teacher_data.get('teachers', [])
-
-    # Calculate pagination
-    total_records = len(teachers)
-    start_idx = (page - 1) * per_page
-    end_idx = start_idx + per_page
-
-    # Get paginated data
-    paginated_teachers = teachers[start_idx:end_idx]
-
-    return jsonify({
-        'status': 'success',
-        'data': paginated_teachers,
-        'pagination': {
-            'page': page,
-            'per_page': per_page,
-            'total_records': total_records,
-            'total_pages': (total_records + per_page - 1) // per_page,
-            'has_next': end_idx < total_records,
-            'has_prev': page > 1
-        },
-        'metadata': {
-            'uploaded_by': teacher_data.get('uploaded_by', 'unknown'),
-            'uploaded_at': teacher_data.get('uploaded_at', 'unknown')
-        }
-    }), 200
-
-
-@recorder_bp.route('/record/<category>', methods=['POST'])
-def record_student(category):
-    """Record a student in a category"""
-    if not require_auth():
-        return jsonify({'error': 'Authentication required'}), 401
-
-    if 'session_id' not in session or session['session_id'] not in session_data:
-        return jsonify({'error': 'No active session'}), 400
-
-    if category not in ['clean', 'dirty', 'red', 'faculty']:
-        return jsonify({'error': 'Invalid category'}), 400
-
-    data = request.get_json(silent=True) or {}
-    input_value = str(data.get('input_value', '') or '').strip()
-    provided_student_id = normalize_name(data.get('student_id'))
-    provided_key_raw = str(data.get('student_key') or '').strip()
-    provided_key = provided_key_raw.lower() if provided_key_raw else ''
-    provided_preferred = normalize_name(data.get('preferred_name') or data.get('preferred'))
-    provided_last = normalize_name(data.get('last_name') or data.get('last'))
-
-    session_id = session['session_id']
-    session_info = session_data[session_id]
-    ensure_session_structure(session_info)
-
-    if category == 'dirty':
-        new_count = session_info.get('dirty_count', 0) + 1
-        session_info['dirty_count'] = new_count
-        record = {
-            'category': 'dirty',
-            'timestamp': datetime.now().isoformat(),
-            'recorded_by': session['user_id'],
-            'display_name': f"Dirty Plate #{new_count}"
-        }
-        session_info['scan_history'].append(record)
-        save_session_data()
-        return jsonify({
-            'status': 'success',
-            'category': 'dirty',
-            'dirty_count': new_count
-        }), 200
-
-    if category == 'faculty':
-        if not input_value:
-            return jsonify({'error': 'Faculty name is required'}), 400
-
-        name_parts = input_value.split()
-        preferred_name = ""
-        last_name = ""
-        if len(name_parts) >= 2:
-            preferred_name = name_parts[0].strip().title()
-            last_name = ' '.join(name_parts[1:]).strip().title()
-        elif len(name_parts) == 1:
-            preferred_name = name_parts[0].strip().title()
-            last_name = ""
-        else:
-            return jsonify({'error': 'Faculty name is required'}), 400
-
-        duplicate_check = any(
-            (record.get('preferred_name') or record.get('first_name', '')).lower() == preferred_name.lower() and
-            record.get('last_name', '').lower() == last_name.lower()
-            for record in session_info['faculty_clean_records']
-        )
-
-        if duplicate_check:
-            return jsonify({
-                'error': 'duplicate',
-                'message': 'Faculty member already recorded in this session'
-            }), 409
-
-        record = {
-            'preferred_name': preferred_name,
-            'first_name': preferred_name,
-            'last_name': last_name,
-            'grade': '',
-            'advisor': '',
-            'house': '',
-            'clan': '',
-            'category': 'faculty',
-            'timestamp': datetime.now().isoformat(),
-            'recorded_by': session['user_id'],
-            'is_manual_entry': True
-        }
-        session_info['faculty_clean_records'].append(record)
-        session_info['scan_history'].append(record)
-        save_session_data()
-        return jsonify({
-            'status': 'success',
-            'preferred_name': preferred_name,
-            'first_name': preferred_name,
-            'last_name': last_name,
-            'category': 'faculty',
-            'is_manual_entry': True,
-            'recorded_by': session['user_id']
-        }), 200
-
-    # Remaining categories rely on student database lookup (clean, red)
-    has_reference = bool(
-        input_value or
-        provided_student_id or
-        provided_key or
-        (provided_preferred and provided_last)
-    )
-    if not has_reference:
-        return jsonify({'error': 'Student ID or Name is required'}), 400
-
-    csv_data = global_csv_data
-    csv_rows = []
-    if isinstance(csv_data, dict):
-        csv_rows = csv_data.get('data') or []
-    if not isinstance(csv_rows, list):
-        csv_rows = []
-
-    student_record = None
-    lookup_profile = None
-    dataset_match = False
-
-    candidate_ids = []
-    if provided_student_id:
-        candidate_ids.append(provided_student_id)
-    key_id = extract_student_id_from_key(provided_key) if provided_key else ''
-    if key_id:
-        candidate_ids.append(key_id)
-    if input_value.isdigit():
-        candidate_ids.append(input_value)
-    else:
-        id_match = re.search(r'\b(\d{3,})\b', input_value)
-        if id_match:
-            candidate_ids.append(id_match.group(1))
-
-    normalized_ids = []
-    for candidate_id in candidate_ids:
-        norm_id = normalize_name(candidate_id)
-        if norm_id and norm_id not in normalized_ids:
-            normalized_ids.append(norm_id)
-
-    for candidate_id in normalized_ids:
-        match = next(
-            (
-                row
-                for row in csv_rows
-                if normalize_name(row.get('Student ID', '')) == candidate_id
-            ),
-            None
-        )
-        if match:
-            student_record = match
-            dataset_match = True
-            break
-
-    cleaned_input = input_value
-    if cleaned_input:
-        cleaned_input = re.sub(r'\([^)]*\)', '', cleaned_input).strip()
-
-    candidate_preferred = provided_preferred
-    candidate_last = provided_last
-
-    if (not candidate_preferred or not candidate_last) and provided_key and not provided_key.startswith('id:'):
-        key_preferred, key_last = split_student_key(provided_key)
-        if key_preferred and not candidate_preferred:
-            candidate_preferred = key_preferred
-        if key_last and not candidate_last:
-            candidate_last = key_last
-
-    if not candidate_preferred or not candidate_last:
-        name_parts = cleaned_input.split()
-        if len(name_parts) >= 2:
-            candidate_preferred = candidate_preferred or name_parts[0]
-            candidate_last = candidate_last or ' '.join(name_parts[1:])
-        elif len(name_parts) == 1:
-            candidate_preferred = candidate_preferred or name_parts[0]
-            candidate_last = candidate_last or ''
-
-    if not student_record and csv_rows and candidate_preferred and candidate_last:
-        input_first = candidate_preferred.lower()
-        input_last = candidate_last.lower()
-        student_record = next(
-            (
-                row
-                for row in csv_rows
-                if str(row.get('Preferred', '')).strip().lower() == input_first
-                and str(row.get('Last', '')).strip().lower() == input_last
-            ),
-            None
-        )
-        if student_record:
-            dataset_match = True
-
-    if not student_record and provided_key and provided_key in student_lookup:
-        lookup_profile = student_lookup.get(provided_key)
-        if lookup_profile:
-            dataset_match = True
-
-    preferred_name = ''
-    last_name = ''
-    grade = ''
-    advisor = ''
-    house = ''
-    clan = ''
-    student_id = ''
-
-    if student_record:
-        preferred_name = str(student_record.get('Preferred', '') or '').strip()
-        last_name = str(student_record.get('Last', '') or '').strip()
-        grade = str(student_record.get('Grade', '') or '').strip()
-        advisor = str(student_record.get('Advisor', '') or '').strip()
-        house = str(student_record.get('House', '') or '').strip()
-        clan = str(student_record.get('Clan', '') or '').strip()
-        student_id = str(student_record.get('Student ID', '') or '').strip()
-    elif lookup_profile:
-        preferred_name = normalize_name(lookup_profile.get('preferred_name'))
-        last_name = normalize_name(lookup_profile.get('last_name'))
-        grade = normalize_name(lookup_profile.get('grade'))
-        advisor = normalize_name(lookup_profile.get('advisor'))
-        house = normalize_name(lookup_profile.get('house'))
-        clan = normalize_name(lookup_profile.get('clan'))
-        student_id = normalize_name(lookup_profile.get('student_id'))
-    else:
-        preferred_name = normalize_name(candidate_preferred or (cleaned_input.split()[0] if cleaned_input else input_value))
-        if cleaned_input:
-            parts = cleaned_input.split()
-            if len(parts) >= 2:
-                candidate_last = candidate_last or ' '.join(parts[1:])
-        last_name = normalize_name(candidate_last)
-
-    if not preferred_name and provided_preferred:
-        preferred_name = provided_preferred
-    if not last_name and provided_last:
-        last_name = provided_last
-    if not student_id and provided_student_id:
-        student_id = provided_student_id
-    if not student_id and provided_key:
-        extracted_id = extract_student_id_from_key(provided_key)
-        if extracted_id:
-            student_id = extracted_id
-    if not student_id and normalized_ids:
-        student_id = normalized_ids[0]
-
-    preferred_name = preferred_name or ''
-    last_name = last_name or ''
-    grade = grade or ''
-    advisor = advisor or ''
-    house = house or ''
-    clan = clan or ''
-    student_id = student_id or ''
-
-    student_key = make_student_key(preferred_name, last_name, student_id) or provided_key or None
-    if student_key:
-        student_key = student_key.lower()
-
-    is_manual_entry = not dataset_match
-
-    student_records = session_info['clean_records'] + session_info['red_records']
-
-    def normalize_field(value):
-        return str(value or '').strip().lower()
-
-    student_key = student_key or ''
-
-    target_student_id = normalize_field(student_id)
-    target_preferred = normalize_field(preferred_name)
-    target_last = normalize_field(last_name)
-    target_grade = normalize_field(grade)
-    target_advisor = normalize_field(advisor)
-    target_house = normalize_field(house)
-    target_clan = normalize_field(clan)
-    target_key = normalize_field(student_key)
-
-    def is_duplicate(existing_record):
-        existing_student_id = normalize_field(existing_record.get('student_id'))
-
-        if target_student_id and existing_student_id:
-            return existing_student_id == target_student_id
-
-        if target_student_id and not existing_student_id:
-            return False
-
-        if not target_student_id and existing_student_id:
-            return False
-
-        existing_key = normalize_field(existing_record.get('student_key'))
-        if target_key and existing_key:
-            return existing_key == target_key
-
-        existing_preferred = normalize_field(existing_record.get('preferred_name') or existing_record.get('first_name'))
-        existing_last = normalize_field(existing_record.get('last_name'))
-        existing_grade = normalize_field(existing_record.get('grade'))
-        existing_advisor = normalize_field(existing_record.get('advisor'))
-        existing_house = normalize_field(existing_record.get('house'))
-        existing_clan = normalize_field(existing_record.get('clan'))
-
-        return (
-            existing_preferred == target_preferred and
-            existing_last == target_last and
-            existing_grade == target_grade and
-            existing_advisor == target_advisor and
-            existing_house == target_house and
-            existing_clan == target_clan
-        )
-
-    duplicate_check = any(
-        is_duplicate(record)
-        for record in student_records
-    )
-
-    if duplicate_check:
-        existing_category = None
-        for cat in ['clean', 'red']:
-            if any(
-                is_duplicate(record)
-                for record in session_info[f'{cat}_records']
-            ):
-                existing_category = cat
-                break
-
-        return jsonify({
-            'error': 'duplicate',
-            'message': f'Student already recorded as {existing_category.upper()} in this session'
-        }), 409
-
-    record = {
-        'preferred_name': preferred_name,
-        'first_name': preferred_name,
-        'last_name': last_name,
-        'grade': grade,
-        'advisor': advisor,
-        'house': house,
-        'clan': clan,
-        'student_id': student_id,
-        'student_key': student_key,
-        'category': category,
-        'timestamp': datetime.now().isoformat(),
-        'recorded_by': session['user_id'],
-        'is_manual_entry': is_manual_entry
-    }
-
-    session_info[f'{category}_records'].append(record)
-    session_info['scan_history'].append(record)
-    save_session_data()
-
-    return jsonify({
-        'status': 'success',
-        'preferred_name': preferred_name,
-        'first_name': preferred_name,
-        'last_name': last_name,
-        'grade': grade,
-        'advisor': advisor,
-        'house': house,
-        'clan': clan,
-        'student_id': student_id,
-        'student_key': student_key,
-        'category': category,
-        'is_manual_entry': is_manual_entry,
-        'recorded_by': session['user_id']
-    }), 200
-
-
-@recorder_bp.route('/session/status', methods=['GET'])
-def get_session_status():
-    """Get current session status with percentage calculations"""
-    if not require_auth_or_guest():
-        return jsonify({'error': 'Authentication or guest access required'}), 401
-
-    if 'session_id' not in session or session['session_id'] not in session_data:
-        return jsonify({'error': 'No active session'}), 400
-
-    session_id = session['session_id']
-    data = session_data[session_id]
-    if is_guest() and not data.get('is_public', True):
-        return jsonify({'error': 'Access denied'}), 403
-
-    ensure_session_structure(data)
-    clean_count = len(data['clean_records'])
-    dirty_count = get_dirty_count(data)
-    red_count = len(data['red_records'])
-    combined_dirty_count = dirty_count + red_count  # Combine dirty + very dirty
-    faculty_clean_count = len(data.get('faculty_clean_records', []))
-    total_recorded = clean_count + dirty_count + red_count
-
-    # Calculate percentages - both clean and faculty clean count as "clean" for ratio
-    combined_clean_count = clean_count + faculty_clean_count
-    total_for_ratio = total_recorded + faculty_clean_count  # Include faculty in total for ratio calculation
-    clean_percentage = (combined_clean_count / total_for_ratio * 100) if total_for_ratio > 0 else 0
-    dirty_percentage = (combined_dirty_count / total_for_ratio * 100) if total_for_ratio > 0 else 0
-
-    return jsonify({
-        'session_id': session_id,
-        'session_name': data['session_name'],
-        'clean_count': clean_count,
-        'dirty_count': dirty_count,
-        'red_count': red_count,
-        'combined_dirty_count': combined_dirty_count,
-        'faculty_clean_count': faculty_clean_count,
-        'total_recorded': total_recorded,
-        'clean_percentage': round(clean_percentage, 1),
-        'dirty_percentage': round(dirty_percentage, 1),
-        'scan_history_count': len(data['scan_history']),
-        'is_discarded': data.get('is_discarded', False),
-        'draw_info': serialize_draw_info(data.get('draw_info', {}))
-    }), 200
-
-
-@recorder_bp.route('/session/history', methods=['GET'])
-def get_session_history():
-    """Get scan history for current session"""
-    if not require_auth_or_guest():
-        return jsonify({'error': 'Authentication or guest access required'}), 401
-
-    if 'session_id' not in session or session['session_id'] not in session_data:
-        return jsonify({'error': 'No active session'}), 400
-
-    session_id = session['session_id']
-    data = session_data[session_id]
-    if is_guest() and not data.get('is_public', True):
-        return jsonify({'error': 'Access denied'}), 403
-
-    return jsonify({
-        'scan_history': data['scan_history']
-    }), 200
-
-
-@recorder_bp.route('/export/csv', methods=['GET'])
-def export_csv():
-    """Export session records as CSV"""
-    if not require_auth():
-        return jsonify({'error': 'Authentication required'}), 401
-
-    if 'session_id' not in session or session['session_id'] not in session_data:
-        return jsonify({'error': 'No active session'}), 400
-
-    session_id = session['session_id']
-    data = session_data[session_id]
-
-    ensure_session_structure(data)
-
-    # Create CSV content with four columns
-    output = io.StringIO()
-
-    # Write header
-    output.write("CLEAN,DIRTY,RED,FACULTY CLEAN\n")
-
-    # Helper to format names using preferred name when available
-    def format_name(record):
-        preferred = (record.get('preferred_name') or record.get('first_name', '') or '').strip()
-        last = (record.get('last_name') or '').strip()
-        full_name = f"{preferred} {last}".strip()
-        return full_name
-
-    # Get all records for each category
-    clean_names = [format_name(record) for record in data['clean_records']]
-    red_names = [format_name(record) for record in data['red_records']]
-    faculty_names = [format_name(record) for record in data.get('faculty_clean_records', [])]
-    dirty_count = get_dirty_count(data)
-
-    # Find the maximum number of records to determine how many rows we need
-    max_records = max(
-        len(clean_names),
-        len(red_names),
-        len(faculty_names),
-        1 if dirty_count > 0 else 0
-    )
-
-    # Write data rows
-    for i in range(max_records):
-        clean_name = clean_names[i] if i < len(clean_names) else ""
-        red_name = red_names[i] if i < len(red_names) else ""
-        faculty_name = faculty_names[i] if i < len(faculty_names) else ""
-        dirty_value = str(dirty_count) if i == 0 and dirty_count > 0 else ""
-
-        output.write(f'"{clean_name}","{dirty_value}","{red_name}","{faculty_name}"\n')
-
-    csv_content = output.getvalue()
-    output.close()
-
-    from flask import Response
-    return Response(
-        csv_content,
-        mimetype='text/csv',
-        headers={'Content-Disposition': f'attachment; filename="{data["session_name"]}_records.csv"'}
-    )
-
-
-@recorder_bp.route('/export/csv/detailed', methods=['GET'])
-def export_detailed_csv():
-    """Export detailed session records without student IDs"""
-    if not require_auth():
-        return jsonify({'error': 'Authentication required'}), 401
-
-    if 'session_id' not in session or session['session_id'] not in session_data:
-        return jsonify({'error': 'No active session'}), 400
-
-    session_id = session['session_id']
-    data = session_data[session_id]
-
-    ensure_session_structure(data)
-
-    detailed_records = []
-    for category in ['clean', 'red']:
-        for record in data.get(f'{category}_records', []):
-            preferred = (record.get('preferred_name') or record.get('first_name', '') or '').strip()
-            last = (record.get('last_name') or '').strip()
-            detailed_records.append({
-                'Category': category.upper(),
-                'Last': last,
-                'Preferred': preferred,
-                'Grade': record.get('grade', ''),
-                'Advisor': record.get('advisor', ''),
-                'House': record.get('house', ''),
-                'Clan': record.get('clan', ''),
-                'Recorded At': record.get('timestamp', ''),
-                'Recorded By': record.get('recorded_by', ''),
-                'Manual Entry': 'Yes' if record.get('is_manual_entry') else 'No'
-            })
-
-    dirty_count = get_dirty_count(data)
-    if dirty_count > 0:
-        detailed_records.append({
-            'Category': 'DIRTY',
-            'Last': '',
-            'Preferred': f'Count: {dirty_count}',
-            'Grade': '',
-            'Advisor': '',
-            'House': '',
-            'Clan': '',
-            'Recorded At': '',
-            'Recorded By': '',
-            'Manual Entry': ''
-        })
-
-    detailed_records.sort(key=lambda x: x['Recorded At'] or '', reverse=True)
-
-    output = io.StringIO()
-    writer = csv.writer(output)
-    header = ['Category', 'Last', 'Preferred', 'Grade', 'Advisor', 'House', 'Clan', 'Recorded At', 'Recorded By', 'Manual Entry']
-    writer.writerow(header)
-    for record in detailed_records:
-        writer.writerow([record[column] for column in header])
-
-    csv_content = output.getvalue()
-    output.close()
-
-    from flask import Response
-    return Response(
-        csv_content,
-        mimetype='text/csv',
-        headers={'Content-Disposition': f'attachment; filename="{data["session_name"]}_detailed_records.csv"'}
-    )
-
-
-@recorder_bp.route('/session/<session_id>/draw/summary', methods=['GET'])
-def get_draw_summary(session_id):
-    """Return draw summary for a session."""
-    if not require_auth_or_guest():
-        return jsonify({'error': 'Authentication or guest access required'}), 401
-
-    if session_id not in session_data:
-        return jsonify({'error': 'Session not found'}), 404
-
-    session_info = session_data[session_id]
-    if is_guest() and not session_info.get('is_public', True):
-        return jsonify({'error': 'Access denied'}), 403
-
-    ensure_session_structure(session_info)
-    summary, _ = get_ticket_summary_for_session(session_id)
-    if not summary:
-        summary = {
-            'session_id': session_id,
-            'session_name': session_info.get('session_name', ''),
-            'created_at': session_info.get('created_at'),
-            'is_discarded': session_info.get('is_discarded', False),
-            'tickets_snapshot': {},
-            'total_tickets': 0.0,
-            'candidates': [],
-            'top_candidates': [],
-            'eligible_count': 0,
-            'excluded_records': 0,
-            'generated_at': datetime.now().isoformat()
-        }
-
-    response = {
-        'session_id': session_id,
-        'session_name': session_info.get('session_name', ''),
-        'created_at': session_info.get('created_at'),
-        'is_discarded': session_info.get('is_discarded', False),
-        'total_tickets': summary.get('total_tickets', 0.0),
-        'eligible_count': summary.get('eligible_count', 0),
-        'excluded_records': summary.get('excluded_records', 0),
-        'top_candidates': summary.get('top_candidates', []),
-        'candidates': summary.get('candidates', []),
-        'ticket_snapshot': summary.get('tickets_snapshot', {}),
-        'generated_at': summary.get('generated_at', datetime.now().isoformat()),
-        'draw_info': serialize_draw_info(session_info.get('draw_info', {}))
-    }
-    return jsonify(response), 200
-
-
-@recorder_bp.route('/session/<session_id>/draw/start', methods=['POST'])
-def start_draw(session_id):
-    """Start a weighted random draw for a session."""
-    if not require_admin():
-        return jsonify({'error': 'Admin access required'}), 403
-
-    if session_id not in session_data:
-        return jsonify({'error': 'Session not found'}), 404
-
-    session_info = session_data[session_id]
-    ensure_session_structure(session_info)
-
-    if session_info.get('is_discarded'):
-        return jsonify({'error': 'Session is discarded from draw calculations'}), 400
-
-    student_records_count = len(session_info.get('clean_records', [])) + len(session_info.get('red_records', []))
-    if student_records_count <= 0:
-        return jsonify({'error': 'No student records available for this session'}), 400
-
-    summary, _ = get_ticket_summary_for_session(session_id)
-    if not summary or summary.get('total_tickets', 0.0) <= 0:
-        return jsonify({'error': 'No eligible tickets available for drawing'}), 400
-
-    tickets_snapshot = summary.get('tickets_snapshot') or {}
-    total_tickets = summary.get('total_tickets', 0.0)
-    profiles = summary.get('profiles') or {}
-
-    cumulative = 0.0
-    target = secure_random.random() * total_tickets
-    chosen_key = None
-    for key, value in tickets_snapshot.items():
-        cumulative += value
-        if target <= cumulative:
-            chosen_key = key
-            break
-    if chosen_key is None and tickets_snapshot:
-        chosen_key = next(iter(tickets_snapshot))
-
-    if chosen_key is None:
-        return jsonify({'error': 'Unable to determine a winner'}), 400
-
-    profile = profiles.get(chosen_key, {}).copy()
-    if not profile:
-        preferred, last = split_student_key(chosen_key)
-        fallback_id = extract_student_id_from_key(chosen_key)
-        profile = {
-            'preferred_name': preferred.title() if preferred else '',
-            'last_name': last.title() if last else '',
-            'grade': '',
-            'advisor': '',
-            'house': '',
-            'clan': '',
-            'student_id': fallback_id
-        }
-    display_name = profile.get('display_name') or format_display_name(profile)
-    winner_tickets = tickets_snapshot.get(chosen_key, 0.0)
-    probability = (winner_tickets / total_tickets * 100.0) if total_tickets > 0 else 0.0
-
-    winner_data = {
-        'key': chosen_key,
-        'preferred_name': profile.get('preferred_name', ''),
-        'last_name': profile.get('last_name', ''),
-        'display_name': display_name,
-        'grade': profile.get('grade', ''),
-        'advisor': profile.get('advisor', ''),
-        'house': profile.get('house', ''),
-        'clan': profile.get('clan', ''),
-        'student_id': profile.get('student_id', ''),
-        'tickets': winner_tickets,
-        'probability': probability
-    }
-
-    timestamp = datetime.now().isoformat()
-    draw_info = session_info['draw_info']
-    draw_info['winner'] = winner_data
-    draw_info['winner_timestamp'] = timestamp
-    draw_info['selected_by'] = session.get('user_id')
-    draw_info['method'] = 'random'
-    draw_info['finalized'] = False
-    draw_info['finalized_at'] = None
-    draw_info['finalized_by'] = None
-    draw_info['override'] = False
-    draw_info['tickets_at_selection'] = total_tickets
-    draw_info['probability_at_selection'] = probability
-    draw_info['eligible_pool_size'] = summary.get('eligible_count', 0)
-
-    history_entry = {
-        'action': 'random_draw',
-        'performed_by': session.get('user_id'),
-        'timestamp': timestamp,
-        'winner_key': chosen_key,
-        'winner_display_name': display_name,
-        'total_tickets': total_tickets,
-        'winner_tickets': winner_tickets,
-        'probability': probability
-    }
-    draw_info['history'].append(history_entry)
-
-    save_session_data()
-
-    updated_summary, _ = get_ticket_summary_for_session(session_id)
-    return jsonify({
-        'status': 'success',
-        'winner': winner_data,
-        'draw_info': serialize_draw_info(draw_info),
-        'summary': updated_summary
-    }), 200
-
-
-@recorder_bp.route('/session/<session_id>/draw/finalize', methods=['POST'])
-def finalize_draw(session_id):
-    """Finalize the current draw winner for a session."""
-    if not require_admin():
-        return jsonify({'error': 'Admin access required'}), 403
-
-    if session_id not in session_data:
-        return jsonify({'error': 'Session not found'}), 404
-
-    session_info = session_data[session_id]
-    ensure_session_structure(session_info)
-    draw_info = session_info.get('draw_info', {})
-    winner = draw_info.get('winner')
-    if not winner:
-        return jsonify({'error': 'No winner to finalize'}), 400
-
-    if not draw_info.get('finalized'):
-        timestamp = datetime.now().isoformat()
-        draw_info['finalized'] = True
-        draw_info['finalized_at'] = timestamp
-        draw_info['finalized_by'] = session.get('user_id')
-        draw_info['history'].append({
-            'action': 'finalized',
-            'performed_by': session.get('user_id'),
-            'timestamp': timestamp,
-            'winner_key': winner.get('key'),
-            'winner_display_name': winner.get('display_name')
-        })
-        save_session_data()
-
-    updated_summary, _ = get_ticket_summary_for_session(session_id)
-    return jsonify({
-        'status': 'success',
-        'finalized': True,
-        'draw_info': serialize_draw_info(draw_info),
-        'summary': updated_summary
-    }), 200
-
-
-@recorder_bp.route('/session/<session_id>/draw/reset', methods=['POST'])
-def reset_draw(session_id):
-    """Reset the draw for a session."""
-    if not require_admin():
-        return jsonify({'error': 'Admin access required'}), 403
-
-    if session_id not in session_data:
-        return jsonify({'error': 'Session not found'}), 404
-
-    session_info = session_data[session_id]
-    ensure_session_structure(session_info)
-    draw_info = session_info.get('draw_info', {})
-    winner = draw_info.get('winner')
-
-    if not winner:
-        return jsonify({'error': 'No draw to reset'}), 400
-
-    if draw_info.get('finalized') and not require_superadmin():
-        return jsonify({'error': 'Only super admins can reset a finalized draw'}), 403
-
-    timestamp = datetime.now().isoformat()
-    draw_info['history'].append({
-        'action': 'reset',
-        'performed_by': session.get('user_id'),
-        'timestamp': timestamp,
-        'previous_winner_key': winner.get('key'),
-        'previous_winner_display_name': winner.get('display_name')
-    })
-
-    draw_info['winner'] = None
-    draw_info['winner_timestamp'] = None
-    draw_info['selected_by'] = None
-    draw_info['method'] = None
-    draw_info['finalized'] = False
-    draw_info['finalized_at'] = None
-    draw_info['finalized_by'] = None
-    draw_info['override'] = False
-    draw_info['tickets_at_selection'] = None
-    draw_info['probability_at_selection'] = None
-    draw_info['eligible_pool_size'] = None
-
-    save_session_data()
-
-    updated_summary, _ = get_ticket_summary_for_session(session_id)
-    return jsonify({
-        'status': 'success',
-        'reset': True,
-        'draw_info': serialize_draw_info(draw_info),
-        'summary': updated_summary
-    }), 200
-
-
-@recorder_bp.route('/session/<session_id>/draw/override', methods=['POST'])
-def override_draw(session_id):
-    """Allow a super admin to override the draw winner."""
-    if not require_superadmin():
-        return jsonify({'error': 'Super admin access required'}), 403
-
-    if session_id not in session_data:
-        return jsonify({'error': 'Session not found'}), 404
-
-    session_info = session_data[session_id]
-    ensure_session_structure(session_info)
-
-    if session_info.get('is_discarded'):
-        return jsonify({'error': 'Session is discarded from draw calculations'}), 400
-
-    student_records_count = len(session_info.get('clean_records', [])) + len(session_info.get('red_records', []))
-    if student_records_count <= 0:
-        return jsonify({'error': 'No student records available for this session'}), 400
-
-    data = request.get_json(silent=True) or {}
-    summary, _ = get_ticket_summary_for_session(session_id)
-    if not summary:
-        return jsonify({'error': 'Unable to load draw summary for this session'}), 400
-
-    candidates = summary.get('candidates') or []
-    tickets_snapshot = summary.get('tickets_snapshot') or {}
-    profiles = summary.get('profiles') or {}
-
-    all_profiles = {}
-    for key, profile in profiles.items():
-        if isinstance(profile, dict):
-            copied = profile.copy()
-            copied.setdefault('key', key)
-            copied.setdefault('display_name', format_display_name(copied))
-            all_profiles[key] = copied
-
-    for record in session_info.get('clean_records', []):
-        profile = build_profile_from_record(record)
-        key = profile.get('key')
-        if key and key not in all_profiles:
-            all_profiles[key] = profile
-
-    for record in session_info.get('red_records', []):
-        profile = build_profile_from_record(record)
-        key = profile.get('key')
-        if key and key not in all_profiles:
-            all_profiles[key] = profile
-
-    for key, info in student_lookup.items():
-        if key not in all_profiles:
-            profile = info.copy()
-            profile['key'] = key
-            profile['display_name'] = format_display_name(profile)
-            all_profiles[key] = profile
-
-    override_key_raw = data.get('student_key')
-    override_key = str(override_key_raw or '').strip().lower()
-    student_id = str(data.get('student_id', '')).strip()
-    input_value = str(data.get('input_value', '')).strip()
-    preferred = data.get('preferred_name')
-    last = data.get('last_name')
-
-    if not override_key and student_id:
-        normalized_id = student_id.lower()
-        match = next(
-            (
-                profile
-                for profile in all_profiles.values()
-                if str(profile.get('student_id', '')).strip().lower() == normalized_id
-            ),
-            None
-        )
-        if match:
-            override_key = match.get('key')
-
-    if not override_key and input_value:
-        normalized_input = input_value.lower()
-        match = next(
-            (
-                profile
-                for profile in all_profiles.values()
-                if (profile.get('display_name') or '').lower() == normalized_input
-            ),
-            None
-        )
-        if not match and input_value.isdigit():
-            match = next(
-                (
-                    profile
-                    for profile in all_profiles.values()
-                    if str(profile.get('student_id', '')).strip() == input_value
-                ),
-                None
-            )
-        if match:
-            override_key = match.get('key')
-
-    if not override_key and (preferred and last or student_id):
-        override_key = make_student_key(preferred, last, student_id)
-
-    if not override_key:
-        return jsonify({'error': 'Unable to determine the override candidate from the provided input'}), 400
-
-    if override_key not in all_profiles:
-        return jsonify({'error': 'Specified student was not found in the uploaded CSV'}), 400
-
-    profile = all_profiles.get(override_key, {}).copy()
-    if not profile:
-        preferred, last = split_student_key(override_key)
-        fallback_id = extract_student_id_from_key(override_key)
-        profile = {
-            'preferred_name': preferred.title() if preferred else '',
-            'last_name': last.title() if last else '',
-            'grade': '',
-            'advisor': '',
-            'house': '',
-            'clan': '',
-            'student_id': fallback_id
-        }
-    display_name = profile.get('display_name') or format_display_name(profile)
-    winner_tickets = tickets_snapshot.get(override_key, 0.0)
-    total_tickets = summary.get('total_tickets', 0.0)
-    probability = (winner_tickets / total_tickets * 100.0) if total_tickets > 0 else 0.0
-
-    winner_data = {
-        'key': override_key,
-        'preferred_name': profile.get('preferred_name', ''),
-        'last_name': profile.get('last_name', ''),
-        'display_name': display_name,
-        'grade': profile.get('grade', ''),
-        'advisor': profile.get('advisor', ''),
-        'house': profile.get('house', ''),
-        'clan': profile.get('clan', ''),
-        'student_id': profile.get('student_id', ''),
-        'tickets': winner_tickets,
-        'probability': probability
-    }
-
-    timestamp = datetime.now().isoformat()
-    draw_info = session_info['draw_info']
-    draw_info['winner'] = winner_data
-    draw_info['winner_timestamp'] = timestamp
-    draw_info['selected_by'] = session.get('user_id')
-    draw_info['method'] = 'random'
-    draw_info['finalized'] = True
-    draw_info['finalized_at'] = timestamp
-    draw_info['finalized_by'] = session.get('user_id')
-    draw_info['override'] = True
-    draw_info['tickets_at_selection'] = summary.get('total_tickets', 0.0)
-    draw_info['probability_at_selection'] = probability
-    draw_info['eligible_pool_size'] = summary.get('eligible_count', 0)
-
-    draw_info['history'].append({
-        'action': 'override',
-        'performed_by': session.get('user_id'),
-        'timestamp': timestamp,
-        'winner_key': override_key,
-        'winner_display_name': display_name,
-        'winner_tickets': winner_tickets,
-        'probability': probability
-    })
-
-    save_session_data()
-
-    updated_summary, _ = get_ticket_summary_for_session(session_id)
-    return jsonify({
-        'status': 'success',
-        'override': True,
-        'winner': winner_data,
-        'draw_info': serialize_draw_info(draw_info),
-        'summary': updated_summary
-    }), 200
-
-
-@recorder_bp.route('/session/<session_id>/draw/discard', methods=['POST'])
-def toggle_discard(session_id):
-    """Toggle whether a session is included in ticket calculations."""
-    if not require_superadmin():
-        return jsonify({'error': 'Super admin access required'}), 403
-
-    if session_id not in session_data:
-        return jsonify({'error': 'Session not found'}), 404
-
-    session_info = session_data[session_id]
-    ensure_session_structure(session_info)
-
-    data = request.get_json(silent=True) or {}
-    discard = bool(data.get('discarded'))
-
-    timestamp = datetime.now().isoformat()
-    message = 'Session state unchanged'
-    if session_info.get('is_discarded') != discard:
-        session_info['is_discarded'] = discard
-        metadata = session_info.get('discard_metadata')
-        if not isinstance(metadata, dict):
-            metadata = {}
-        if discard:
-            metadata['discarded_by'] = session.get('user_id')
-            metadata['discarded_at'] = timestamp
-            message = 'Session discarded from draw calculations'
-        else:
-            metadata['restored_by'] = session.get('user_id')
-            metadata['restored_at'] = timestamp
-            message = 'Session reinstated for draw calculations'
-        session_info['discard_metadata'] = metadata
-        session_info['draw_info']['history'].append({
-            'action': 'session_discarded' if discard else 'session_restored',
-            'performed_by': session.get('user_id'),
-            'timestamp': timestamp
-        })
-        save_session_data()
-
-    summary, _ = get_ticket_summary_for_session(session_id)
-    return jsonify({
-        'status': 'success',
-        'discarded': session_info.get('is_discarded', False),
-        'message': message,
-        'draw_info': serialize_draw_info(session_info.get('draw_info', {})),
-        'summary': summary,
-        'discard_metadata': session_info.get('discard_metadata', {})
-    }), 200
-
-
-@recorder_bp.route('/admin/overview', methods=['GET'])
-def admin_overview():
-    """Get admin overview data"""
-    if not require_admin():
-        return jsonify({'error': 'Admin access required'}), 403
-
-    # Get all users
-    users = []
-    for user in list_all_users():
-        users.append({
-            'username': user['username'],
-            'name': user['name'],
-            'role': user['role']
-        })
-
-    # Get all sessions
-    sessions = []
-    for session_id, session_info in session_data.items():
-        ensure_session_structure(session_info)
-        total_records = len(session_info['clean_records']) + get_dirty_count(session_info) + len(session_info['red_records'])
-        sessions.append({
-            'session_id': session_id,
-            'session_name': session_info['session_name'],
-            'owner': session_info['owner'],
-            'total_records': total_records,
-            'created_at': session_info['created_at'],
-            'faculty_clean_count': len(session_info.get('faculty_clean_records', [])),
-            'is_discarded': session_info.get('is_discarded', False),
-            'draw_info': serialize_draw_info(session_info.get('draw_info', {}))
-        })
-
-    return jsonify({
-        'users': users,
-        'sessions': sessions
-    }), 200
-
-
-@recorder_bp.route('/session/scan-history', methods=['GET'])
-def get_scan_history():
-    """Get scan history for current session"""
-    if not require_auth_or_guest():
-        return jsonify({'error': 'Authentication or guest access required'}), 401
-
-    if 'session_id' not in session or session['session_id'] not in session_data:
-        return jsonify({'error': 'No active session'}), 400
-
-    session_id = session['session_id']
-    data = session_data[session_id]
-    ensure_session_structure(data)
-
-    # Format scan history for display
-    formatted_history = []
-    for record in data['scan_history']:
-        name = (record.get('display_name') or '').strip()
-        if not name:
-            preferred = (record.get('preferred_name') or record.get('first_name', '') or '').strip()
-            last = (record.get('last_name') or '').strip()
-            name = f"{preferred} {last}".strip()
-
-        category = record.get('category', '').lower()
-        if not name:
-            if category == 'dirty':
-                name = 'Dirty Plate'
-            elif category == 'faculty':
-                name = 'Faculty Clean Plate'
-
-        formatted_history.append({
-            'timestamp': record.get('timestamp'),
-            'name': name,
-            'category': category.upper(),
-            'is_manual_entry': record.get('is_manual_entry', False)
-        })
-
-    # Sort by timestamp (most recent first)
-    formatted_history.sort(key=lambda x: x['timestamp'], reverse=True)
-
-    return jsonify({
-        'scan_history': formatted_history
-    }), 200
-
-
-@recorder_bp.route('/superadmin/change-role', methods=['POST'])
-def change_user_role():
-    """Change user role (super admin only)"""
-    if 'user_id' not in session:
-        return jsonify({'error': 'Authentication required'}), 401
-
-    current_user = get_current_user()
-    if not current_user or current_user['role'] != 'superadmin':
-        return jsonify({'error': 'Super admin access required'}), 403
-
-    data = request.get_json()
-    target_username = data.get('username')
-    new_role = data.get('role')
-
-    if not target_username or not new_role:
-        return jsonify({'error': 'Username and role are required'}), 400
-
-    if new_role not in ['user', 'admin', 'superadmin']:
-        return jsonify({'error': 'Invalid role'}), 400
-
-    # Cannot change own role
-    if target_username == session['user_id']:
-        return jsonify({'error': 'Cannot change your own role'}), 400
-
-    target_user_model = get_user_by_username(target_username)
-    if not target_user_model:
-        return jsonify({'error': 'User not found'}), 404
-
-    try:
-        update_user_credentials(target_user_model, role=new_role)
-    except Exception:
-        return jsonify({'error': 'Failed to update user role'}), 500
-
-    return jsonify({
-        'status': 'success',
-        'message': f'User role changed to {new_role} successfully'
-    }), 200
-
-
-@recorder_bp.route('/superadmin/delete-account', methods=['POST'])
-def delete_user_account():
-    """Delete user account (super admin only)"""
-    if 'user_id' not in session:
-        return jsonify({'error': 'Authentication required'}), 401
-
-    current_user = get_current_user()
-    if not current_user or current_user['role'] != 'superadmin':
-        return jsonify({'error': 'Super admin access required'}), 403
-
-    data = request.get_json()
-    target_username = data.get('username')
-
-    if not target_username:
-        return jsonify({'error': 'Username is required'}), 400
-
-    # Cannot delete own account
-    if target_username == session['user_id']:
-        return jsonify({'error': 'Cannot delete your own account'}), 400
-
-    target_user_model = get_user_by_username(target_username)
-    if not target_user_model:
-        return jsonify({'error': 'User not found'}), 404
-
-    # Delete user account
-    try:
-        db_session.delete(target_user_model)
-        db_session.commit()
-    except Exception:
-        db_session.rollback()
-        return jsonify({'error': 'Failed to delete user account'}), 500
-
-    # Clean up user's CSV data file (legacy compatibility)
-    user_csv_file = f"user_csv_{target_username}.json"
-    if os.path.exists(user_csv_file):
-        os.remove(user_csv_file)
-
-    # Remove user's sessions
-    sessions_to_remove = []
-    for session_id, session_info in session_data.items():
-        if session_info.get('owner') == target_username:
-            sessions_to_remove.append(session_id)
-
-    for session_id in sessions_to_remove:
-        del session_data[session_id]
-    save_session_data()
-
-    return jsonify({
-        'status': 'success',
-        'message': 'User account deleted successfully'
-    }), 200
-
-
-@recorder_bp.teardown_app_request
-def shutdown_session(exception=None):
-    """Ensure the scoped session is removed after each request."""
-    db_session.remove()
